DimensionalData.jl

DimensionalData.DimensionalData — Module

DimensionalData.jl provides types and methods for indexing with named dimensions, using named dimensions in Base and Statistics methods instead of Integer dims, and selecting data from dimension values instead of using indices directly.

Dimensions

Dimensions are simply types that wrap values. They both store dimension values and are used for dimension lookup or indices, ranges or dimension number. X, Y, Z and Time are the unexported defaults, add this line to use them:

using DimensionalData: X, Y, Z, Time

A generalised Dim type is available to use arbitrary symbols to name dimensions. Custom dimensions can be defined using the @dim macro.

We can use dim wrappers for indexing, so that the dimension order in the underlying array does not need to be known:

a[X(1:10), Y(1:4)]

Selectors

Selectors find indices in the dimension based on values At, Near, or Between the index value(s).

We can use selectors in conjuction with dim wrappers:

a[X(1:10), Y<|At(25.7)]

Without dim wrappers selectors must be in the right order:

usin Unitful
a[Near(23u"s"), Between(10.5u"m", 50.5u"m")]

Dim types or objects can be used instead of a dimension number in many Base and Statistics methods:

mean(a; dims=X)
std(a; dims=Y())

DimensionalData.AbstractAlignedGrid — Type

A grid dimension aligned exactly with a standard dimension, such as lattitude or longitude.

DimensionalData.AbstractDimension — Type

An AbstractDimension tags the dimensions in an AbstractArray.

It can also contain spatial coordinates and their metadata. For simplicity, the same types are used both for storing dimension information and for indexing.

DimensionalData.AlignedGrid — Type

An alligned grid without known regular spacing. These grids will generally be paired with a vector of coordinates along the dimension, instead of a range.

As the size of the cells is not known, the bounds must be actively tracked.

Fields

order: Order trait indicating array and index order

locus: Locus trait indicating the position of the indexed point within the cell span

sampling: Sampling trait indicating wether the grid cells are single samples or means

bounds: the outer edges of the grid (different to the first and last coordinate).

DimensionalData.At — Type

At(x)

Selector that exactly matches the value on the passed-in dimensions, or throws an error. For ranges and arrays, every value must match an existing value - not just the end points.

DimensionalData.Between — Type

Between(a, b)

Selector that retreive all indices located between 2 values.

DimensionalData.CategoricalGrid — Type

A grid dimension where the values are categories.

Fields

order: Order trait indicating array and index order

DimensionalData.Center — Type

Indicates dimensions that are defined by their center coordinates/time/position.

DimensionalData.DependentGrid — Type

Traits describing a grid dimension that is dependent on other grid dimensions.

Indexing into a dependent dimension must provide all other dependent dimensions.

DimensionalData.Dim — Type

A generic dimension. For use when custom dims are required when loading data from a file. The sintax is ugly and verbose to use for indexing, ie Dim{:lat}(1:9) rather than Lat(1:9). This is the main reason they are not the only type of dimension availabile.

DimensionalData.DimensionalArray — Type

DimensionalArray(A::AbstractArray, dims::Tuple, refdims::Tuple)

A basic DimensionalArray type.

Maintains and updates its dimensions through transformations

DimensionalData.DimensionalArray — Method

DimensionalArray(A::AbstractArray, dims::Tuple; refdims=())

Constructor with optional refdims keyword.

Example:

using Dates, DimensionalData
using DimensionalData: Time, X
timespan = DateTime(2001):Month(1):DateTime(2001,12)
A = DimensionalArray(rand(12,10), (Time(timespan), X(10:10:100))) 
A[X<|Near([12, 35]), Time<|At(DateTime(2001,5))]
A[Near(DateTime(2001, 5, 4)), Between(20, 50)]

DimensionalData.End — Type

Indicates dimensions that are defined by their end coordinates/time/position.

DimensionalData.Grid — Type

Traits describing the grid type of a dimension.

DimensionalData.IndependentGrid — Type

A grid dimension that is independent of other grid dimensions.

DimensionalData.Locus — Type

Indicate the position of index values in grid cells.

This is frequently Start for time series, but may be Center for spatial data.

DimensionalData.LookupGrid — Type

A grid dimension that uses an array lookup to convert dimension from dim(grid) to dims(array).

Fields

dims: a tuple containing dimenension types or symbols matching the order needed to index the lookup matrix.

sampling: a Sampling trait indicating wether the grid cells are sampled points or means

DimensionalData.Metadata — Type

Abstract type for dimension metadata wrappers.

DimensionalData.MultiSample — Type

Multiple samples from the span combined using method M, where M is typeof(mean), typeof(sum) etc.

DimensionalData.Near — Type

Near(x)

Selector that selects the nearest index to its contained value(s)

DimensionalData.Ordered — Type

Trait container for dimension and array ordering in AlignedGrid.

The default is Ordered(Forward(), Forward())

All combinations of forward and reverse order for data and indices seem to occurr in real datasets, as strange as that seems. We cover these possibilities by specifying the order of both explicitly.

Knowing the order of indices is important for using methods like searchsortedfirst() to find indices in sorted lists of values. Knowing the order of the data is then required to map to the actual indices. It's also used to plot the data later - which always happens in smallest to largest order.

Base also defines Forward and Reverse, but they seem overly complicated for our purposes.

DimensionalData.RegularGrid — Type

An alligned grid known to have equal spacing between all cells.

Fields

order: Order trait indicating array and index order

locus: Locus trait indicating the position of the indexed point within the cell span

sampling: Sampling trait indicating wether the grid cells are single samples or means

span: the size of a grid step, such as 1u"km" or Month(1)

DimensionalData.Sampling — Type

Indicates wether the cell value is specific to the locus point or is related to the whole the span.

The span may contain a value if the distance between locii if known. This will often be identical to the distance between any two sequential cell values, but may be distinct due to rounding errors in a vector index, or context-dependent spans such as Month.

DimensionalData.Selector — Type

Selectors indicate that index values are not indices, but points to be selected from the dimension values, such as DateTime objects on a Time dimension.

DimensionalData.SingleSample — Type

Each cell value represents a single discrete sample taken at the index location.

DimensionalData.Start — Type

Indicates dimensions that are defined by their start coordinates/time/position.

DimensionalData.TransformedGrid — Type

Grid type using an affine transformation to convert dimension from dim(grid) to dims(array).

Fields

dims: a tuple containing dimenension types or symbols matching the order needed by the transform function.

sampling: a Sampling trait indicating wether the grid cells are sampled points or means

DimensionalData.UnknownGrid — Type

Fallback grid type

DimensionalData.UnknownSampling — Type

The sampling method is unknown.

DimensionalData.Unordered — Type

Trait indicating that the array or dimension has no order.

DimensionalData.dims — Function

dims(x)

Return a tuple of the dimensions for a dataset. These can contain the coordinate ranges if bounds() and select() are to be used, or you want them to be shown on plots in place of the array indices.

They can also contain a units string or unitful unit to use and plot dimension units.

This is the only method required for this package to work. It probably requires defining a dims field on your object to store dims in.

DimensionalData.dims — Method

Get the dimension(s) matching the type(s) of the lookup dimension.

DimensionalData.label — Function

label(x)

Get a plot label for data or a dimension. This will include the name and units if they exist, and anything else that should be shown on a plot.

DimensionalData.metadata — Function

metadata(x)

Return the metadata of a dimension or data object.

DimensionalData.name — Function

name(x)

Get the name of data or a dimension.

DimensionalData.refdims — Function

refdims(x)

Reference dimensions for an array that is a slice or view of another array with more dimensions.

slicedims(a, dims) returns a tuple containing the current new dimensions and the new reference dimensions. Refdims can be stored in a field or disgarded, as it is mostly to give context to plots. Ignoring refdims will simply leave some captions empty.

DimensionalData.shortname — Function

shortname(x)

Get the short name of array data or a dimension.

DimensionalData.units — Function

units(x)

Return the units of a dimension. This could be a string, a unitful unit, or nothing.

DimensionalData.val — Function

val(x)

Return the contained value of a wrapper object, otherwise just returns the object.

DimensionalData.AbstractParametricDimension — Type

Dimensions with user-set type paremeters

DimensionalData.Forward — Type

Trait indicating that the array or dimension is in the normal forward order.

DimensionalData.Reverse — Type

Trait indicating that the array or dimension is in the reverse order. Selector lookup or plotting will be reversed.

Base.permutedims — Method

Sort dimensions into the order they take in the array.

Missing dimensions are replaced with nothing

DimensionalData.dimnum — Method

Get the number of an AbstractDimension as ordered in the array

DimensionalData.dims2indices — Function

Convert a tuple of AbstractDimension to indices, ranges or Colon.

DimensionalData.formatdims — Method

Format the dimension to match internal standards.

Mostily this means converting tuples and UnitRanges to LinRange, which is easier to handle. Errors are thrown if dims don't match the array dims or size.

DimensionalData.rebuild — Function

rebuild(x::AbstractDimensionalArray, data, [dims], [refdims])
rebuild(x::AbstractDimension, val, [grid], [metadata])
rebuild(x; kwargs...)

Rebuild an object struct with updated values.

DimensionalData.reducedims — Method

Replace the specified dimensions with an index of length 1 to match a new array size where the dimension has been reduced to a length of 1, but the number of dimensions has not changed.

Used in mean, reduce, etc.

Grid traits are also updated to correspond to the change in cell span, sampling type and order.

DimensionalData.slicedims — Method

Slice the dimensions to match the axis values of the new array

All methods returns a tuple conatining two tuples: the new dimensions, and the reference dimensions. The ref dimensions are no longer used in the new struct but are useful to give context to plots.

Called at the array level the returned tuple will also include the previous reference dims attached to the array.

DimensionalData.@dim — Macro

@dim typ [name=string(typ)] [shortname=string(typ)]

Macro to easily define specific dimensions.

Example:

@dim Lat "Lattitude"