API

To use the functionality of DimensionalData in your module, dispatch on AbstractDimensionalArray and AbstractDimension.

Core types

Arrays:

DimensionalData.AbstractDimensionalArray — Type

Parent type for all dimensional arrays.

DimensionalData.DimensionalArray — Type

DimensionalArray(data, dims, refdims, name)

The main subtype of AbstractDimensionalArray. Maintains and updates its dimensions through transformations and moves dimensions to refdims after reducing operations (like e.g. mean).

Dimensions:

Missing docstring.

Missing docstring for AbstractDimension. Check Documenter's build log for details.

DimensionalData.DependentDim — Type

Abstract supertype for Dependent dimensions. Will plot on the Y axis.

DimensionalData.IndependentDim — Type

Abstract supertype for independent dimensions. Will plot on the X axis.

Missing docstring.

Missing docstring for CategoricalDim. Check Documenter's build log for details.

DimensionalData.XDim — Type

Abstract parent type for all X dimensions.

DimensionalData.YDim — Type

Abstract parent type for all Y dimensions.

DimensionalData.ZDim — Type

Abstract parent type for all Z dimensions.

DimensionalData.TimeDim — Type

Abstract parent type for all time dimensions.

DimensionalData.X — Type

X dimension. `X <: XDim <: IndependentDim

DimensionalData.Y — Type

Y dimension. `Y <: YDim <: DependentDim

DimensionalData.Z — Type

Z dimension. `Z <: ZDim <: Dimension

DimensionalData.Ti — Type

Time dimension. `Ti <: TimeDim <: IndependentDim

Time is already used by Dates, so we use Ti to avoid clashing.

DimensionalData.Dim — Type

A generic dimension. For use when custom dims are required when loading data from a file. The sintax is ugly and verbose to use for indexing, ie Dim{:lat}(1:9) rather than Lat(1:9). This is the main reason they are not the only type of dimension availabile.

DimensionalData.@dim — Macro

@dim typ [supertype=Dimension] [name=string(typ)] [shortname=string(typ)]

Macro to easily define specific dimensions.

Example:

@dim Lat "Lattitude" "lat"
@dim Lon XDim "Longitude"

Getting basic info

These useful functions for obtaining information from your dimensional data:

DimensionalData.dims — Function

dims(x)

Return a tuple of the dimensions for a dataset. These can contain the coordinate ranges if bounds() and select() are to be used, or you want them to be shown on plots in place of the array indices.

They can also contain a units string or unitful unit to use and plot dimension units.

This is the only method required for this package to work. It probably requires defining a dims field on your object to store dims in.

DimensionalData.hasdim — Function

hasdim(A, lookup)

Check if an object or tuple contains an Dimension, or a tuple of dimensions.

DimensionalData.dimnum — Function

dimnum(A, lookup)

Get the number(s) of Dimension(s) as ordered in the dimensions of an object.

DimensionalData.name — Function

name(x)

Get the name of data or a dimension.

DimensionalData.val — Function

val(x)

Return the contained value of a wrapper object, otherwise just returns the object.

As well as others related to obtained metadata:

DimensionalData.bounds — Function

bounds(x, [dims])

Return the bounds of all dimensions of an object, of a specific dimension, or of a tuple of dimensions.

Bounds are allways return in ascending order.

DimensionalData.label — Function

label(x)

Get a plot label for data or a dimension. This will include the name and units if they exist, and anything else that should be shown on a plot.

DimensionalData.metadata — Function

metadata(x)

Return the metadata of a dimension or data object.

DimensionalData.refdims — Function

refdims(x)

Reference dimensions for an array that is a slice or view of another array with more dimensions.

slicedims(a, dims) returns a tuple containing the current new dimensions and the new reference dimensions. Refdims can be stored in a field or disgarded, as it is mostly to give context to plots. Ignoring refdims will simply leave some captions empty.

DimensionalData.shortname — Function

shortname(x)

Get the short name of array data or a dimension.

DimensionalData.units — Function

units(x)

Return the units of a dimension. This could be a string, a unitful unit, or nothing.

DimensionalData.data — Function

data(x)

Return the data wrapped by the dimentional array. This may not be the same as Base.parent, as it should never include data outside the bounds of the dimensions.

In a disk based AbstractDimensionalArray, data may need to load data from disk.

Selectors

DimensionalData.Selector — Type

Selectors indicate that index values are not indices, but points to be selected from the dimension values, such as DateTime objects on a Time dimension.

DimensionalData.At — Type

At(x)

Selector that exactly matches the value on the passed-in dimensions, or throws an error. For ranges and arrays, every value must match an existing value - not just the end points.

DimensionalData.Near — Type

Near(x)

Selector that selects the nearest index. With PointSampling this is simply the index nearest to the contained value, however with `IntervalSampling it is the interval center nearest to the contained value. This will be offset from the index value for Start and End loci.

DimensionalData.Between — Type

Between(a, b)

Selector that retreive all indices located between 2 values.

For IntervalSampling the whole interval must be lie between the values. For PointSampling the points must fall between the 2 values. These different sampling traits will often give different results with the same index and values - this is the intended behaviour.

Grids

DimensionalData.Grid — Type

Traits describing the grid type of a dimension.

Missing docstring.

Missing docstring for IndependentGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AlignedGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BoundedGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularGrid. Check Documenter's build log for details.

DimensionalData.CategoricalGrid — Type

A grid dimension where the values are categories.

Fields

order: Order trait indicating array and index order

DimensionalData.UnknownGrid — Type

Unkwown grid type. Will be converted automatically to another grid type when possible.

Missing docstring.

Missing docstring for DependentGrid. Check Documenter's build log for details.

DimensionalData.TransformedGrid — Type

Grid trait that uses an affine transformation to convert dimensions from dims(grid) to dims(array).

Fields

dims: a tuple containing dimenension types or symbols matching the order needed by the transform function.

Tracking the order of arrays and indices:

DimensionalData.Unordered — Type

Trait indicating that the array or dimension has no order.

DimensionalData.Ordered — Type

Trait container for dimension and array ordering.

The default is Ordered(Forward(), Forward(), Forward())

All combinations of forward and reverse order for data and indices seem to occurr in real datasets, as strange as that seems. We cover these possibilities by specifying the order of both explicitly.

Knowing the order of indices is important for using methods like searchsortedfirst() to find indices in sorted lists of values. Knowing the order of the data is then required to map to the actual indices. It's also used to plot the data later - which always happens in smallest to largest order.

Base also defines Forward and Reverse, but they seem overly complicated for our purposes.

Loci

DimensionalData.Locus — Type

Locii indicate the position of index values in grid cells.

Locii are often Start for time series, but often Center for spatial data.

DimensionalData.Center — Type

Indicates dimension index that matches the center coordinates/time/position.

DimensionalData.Start — Type

Indicates dimension index that matches the start coordinates/time/position.

DimensionalData.End — Type

Indicates dimension index that matches the end coordinates/time/position.

DimensionalData.UnknownLocus — Type

Indicates dimension where the index position is not known.

Low-level API

DimensionalData.rebuild — Function

rebuild(x::AbstractDimensionalArray, data, [dims], [refdims], [name])
rebuild(x::AbstractDimensionalArray, data, [name])
rebuild(x::AbstractDimension, val, [grid], [metadata])
rebuild(x; kwargs...)

Rebuild an object struct with updated values.

DimensionalData.formatdims — Function

formatdims(A, dims)

Format the passed-in dimension(s).

Mostily this means converting indexes of tuples and UnitRanges to LinRange, which is easier to handle internally. Errors are also thrown if dims don't match the array dims or size.

If a Grid hasn't been specified, a grid type is chosen based on the type and element type of the index:

DimensionalData.reducedims — Function

Replace the specified dimensions with an index of length 1 to match a new array size where the dimension has been reduced to a length of 1, but the number of dimensions has not changed.

Used in mean, reduce, etc.

Grid traits are also updated to correspond to the change in cell step, sampling type and order.

DimensionalData.slicedims — Function

Slice the dimensions to match the axis values of the new array

All methods returns a tuple conatining two tuples: the new dimensions, and the reference dimensions. The ref dimensions are no longer used in the new struct but are useful to give context to plots.

Called at the array level the returned tuple will also include the previous reference dims attached to the array.