API

To use the functionality of DimensionalData in your module, dispatch on AbstractDimArray and AbstractDimension.

Arrays

DimensionalData.AbstractDimArray — Type

AbstracDimArray <: AbstractArray

Abstract supertype for all "dim" arrays.

These arrays return a Tuple of Dimension from a dims method, and can be rebuilt using rebuild.

parent must return the source array.

They should have metadata, name and refdims methods, although these are optional.

A rebuild method for AbstractDimArray must accept data, dims, refdims, name, metadata arguments.

Indexing AbstractDimArray with non-range AbstractArray has undefined effects on the Dimension index. Use forward-ordered arrays only"

DimensionalData.DimArray — Type

DimArray <: AbstractDimArray

DimArray(data, dims, refdims, name)
DimArray(data, dims::Tuple [, name::Symbol]; refdims=(), metadata=NoMetadata())

The main concrete subtype of AbstractDimArray.

DimArray maintains and updates its Dimensions through transformations and moves dimensions to reference dimension refdims after reducing operations (like e.g. mean).

Arguments/Fields

data: An AbstractArray.

dims: A Tuple of Dimension

name: A string name for the array. Shows in plots and tables.

refdims: refence dimensions. Usually set programmatically to track past slices and reductions of dimension for labelling and reconstruction.

metadata: Array metadata, or NoMetadata()

Indexing can be done with all regular indices, or with Dimensions and/or Selectors. Indexing AbstractDimArray with non-range AbstractArray has undefined effects on the Dimension index. Use forward-ordered arrays only"

Example:

using Dates, DimensionalData

ti = (Ti(DateTime(2001):Month(1):DateTime(2001,12)),
x = X(10:10:100))
A = DimArray(rand(12,10), (ti, x), "example")

julia> A[X(Near([12, 35])), Ti(At(DateTime(2001,5)))];

julia> A[Near(DateTime(2001, 5, 4)), Between(20, 50)];

Multi-array datasets

DimensionalData.AbstractDimStack — Type

AbstractDimStack

Abstract supertype for dimensional stacks.

These have multiple layers of data, but share dimensions.

DimensionalData.DimStack — Type

DimStack <: AbstractDimStack

DimStack(data::AbstractDimArray...)
DimStack(data::Tuple{Vararg{<:AbstractDimArray}})
DimStack(data::NamedTuple{Keys,Vararg{<:AbstractDimArray}})
DimStack(data::NamedTuple, dims::DimTuple; metadata=NoMetadata())

DimStack holds multiple objects sharing some dimensions, in a NamedTuple. Indexing operates as for AbstractDimArray, except it occurs for all data layers of the stack simulataneously. Layer objects can hold values of any type.

DimStack can be constructed from multiple AbstractDimArray or a NamedTuple of AbstractArray and a matching dims Tuple. If AbstractDimArrays have the same name they will be given the name :layer1, substitiuting the layer number for 1.

getindex with Int or Dimensions or Selectors that resolve to Int will return a NamedTuple of values from each layer in the stack. This has very good performace, and usually takes less time than the sum of indexing each array separately.

Indexing with a Vector or Colon will return another DimStack where all data layers have been sliced. setindex! must pass a Tuple or NamedTuple maching the layers.

Most Base and Statistics methods that apply to AbstractArray can be used on all layers of the stack simulataneously. The result is a DimStack, or a NamedTuple if methods like mean are used without dims arguments, and return a single non-array value.

Example

julia> using DimensionalData

julia> A = [1.0 2.0 3.0; 4.0 5.0 6.0];

julia> dimz = (X([:a, :b]), Y(10.0:10.0:30.0))
(X{Vector{Symbol}, AutoMode{AutoOrder}, NoMetadata}([:a, :b], AutoMode{AutoOrder}(AutoOrder()), NoMetadata()), Y{StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}}, AutoMode{AutoOrder}, NoMetadata}(10.0:10.0:30.0, AutoMode{AutoOrder}(AutoOrder()), NoMetadata()))

julia> da1 = DimArray(1A, dimz, :one);

julia> da2 = DimArray(2A, dimz, :two);

julia> da3 = DimArray(3A, dimz, :three);

julia> s = DimStack(da1, da2, da3);

julia> s[:b, 10.0]
(one = 4.0, two = 8.0, three = 12.0)

julia> s[X(:a)] isa DimStack
true

Core types

Dimensions:

DimensionalData.Dimension — Type

Dimension

Abstract supertype of all dimension types.

Example concrete implementations are X, Y, Z, Ti (Time), and the custom [Dim]@ref) dimension.

Dimensions label the axes of an AbstractDimArray, or other dimensional objects, and are used to index into the array.

They may also provide an alternate index to lookup for each array axis. This may be any AbstractVector matching the array axis length, or a Val holding a tuple for compile-time index lookups.

Dimensions also have mode and metadata fields.

mode gives more details about the dimension, such as that it is Categorical or Sampled as Points or Intervals along some transect. DimensionalData will attempt to guess the mode from the passed-in index value.

metadata can hold any metadata object adding more information about the array axis - useful for extending DimensionalData for specific contexts, like geospatial data in GeoData.jl. By default it is NoMetadata().

Example:

using DimensionalData, Dates

x = X(2:2:10)
y = Y(['a', 'b', 'c'])
ti = Ti(DateTime(2021, 1):Month(1):DateTime(2021, 12))

A = DimArray(zeros(3, 5, 12), (y, x, ti))

# output

3×5×12 DimArray{Float64,3} with dimensions:
  Y: Char[a, b, c] (Categorical - Unordered)
  X: 2:2:10 (Sampled - Ordered Regular Points)
  Ti (Time): DateTime("2021-01-01T00:00:00"):Month(1):DateTime("2021-12-01T00:00:00") (Sampled - Ordered Regular Points)
[:, :, 1]
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
[and 11 more slices...]

For simplicity, the same Dimension types are also used as wrappers in getindex, like:

x = A[X(2), Y(3)]

# output

12-element DimArray{Float64,1} with dimensions:
  Ti (Time): DateTime("2021-01-01T00:00:00"):Month(1):DateTime("2021-12-01T00:00:00") (Sampled - Ordered Regular Points)
and reference dimensions:
  Y(c): c (Categorical - Unordered)
  X(4): 4 (Sampled - Ordered Regular Points)
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0

A Dimension can also wrap Selector.

x = A[X(Between(3, 4)), Y(At('b'))]

# output

1×12 DimArray{Float64,2} with dimensions:
  X: 4:2:4 (Sampled - Ordered Regular Points)
  Ti (Time): DateTime("2021-01-01T00:00:00"):Month(1):DateTime("2021-12-01T00:00:00") (Sampled - Ordered Regular Points)
and reference dimensions:
  Y(b): b (Categorical - Unordered)
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

Dimension objects may have mode and metadata fields to track additional information about the data and the index, and their relationship.

DimensionalData.DependentDim — Type

DependentDim <: Dimension

Abstract supertype for Dependent dimensions. These will plot on the Y axis.

DimensionalData.IndependentDim — Type

IndependentDim <: Dimension

Abstract supertype for independent dimensions. Thise will plot on the X axis.

DimensionalData.XDim — Type

XDim <: IndependentDim

Abstract supertype for all X dimensions.

DimensionalData.YDim — Type

YDim <: DependentDim

Abstract supertype for all Y dimensions.

DimensionalData.ZDim — Type

ZDim <: DependentDim

Abstract supertype for all Z dimensions.

DimensionalData.TimeDim — Type

TimeDim <: IndependentDim

Abstract supertype for all time dimensions.

In a TimeDime with Interval sampling the locus will automatically be set to Start(). Dates and times generally refer to the start of a month, hour, second etc., not the central point as is more common with spatial data. `

DimensionalData.X — Type

X <: XDim

X(val=:; mode=AutoMode(), metadata=NoMetadata())

X Dimension. X <: XDim <: IndependentDim

Example:

xdim = X(2:2:10)
# Or
val = A[X(1)]
# Or
mean(A; dims=X)

DimensionalData.Y — Type

Y <: YDim

Y(val=:; mode=AutoMode(), metadata=NoMetadata())

Y Dimension. Y <: YDim <: DependentDim

Example:

ydim = Y(['a', 'b', 'c'])
# Or
val = A[Y(1)]
# Or
mean(A; dims=Y)

DimensionalData.Z — Type

Z <: ZDim

Z(val=:; mode=AutoMode(), metadata=NoMetadata())

Z Dimension. Z <: ZDim <: Dimension

Example:

zdim = Z(10:10:100)
# Or
val = A[Z(1)]
# Or
mean(A; dims=Z)

DimensionalData.Ti — Type

m Ti <: TimeDim

Ti(val=:; mode=AutoMode(), metadata=NoMetadata())

Time Dimension. Ti <: TimeDim <: IndependentDim

Time is already used by Dates, and T is a common type parameter, We use Ti to avoid clashes.

Example:

timedim = Ti(DateTime(2021, 1):Month(1):DateTime(2021, 12))
# Or
val = A[Ti(1)]
# Or
mean(A; dims=Ti)

DimensionalData.ParametricDimension — Type

Abstract supertype for Dimensions with user-set type paremeters

DimensionalData.Dim — Type

Dim{S}()
Dim{S}(val=:; mode=AutoMode(), metadata=NoMetadata())
Dim{S}(val, mode, metadata=NoMetadata())

A generic dimension. For use when custom dims are required when loading data from a file. Can be used as keyword arguments for indexing.

Dimension types take precedence over same named Dim types when indexing with symbols, or e.g. creating Tables.jl keys.

using DimensionalData

dim = Dim{:custom}(['a', 'b', 'c'])

# output

Dim{:custom}:
  val: Char[a, b, c]
Dim{:custom, Vector{Char}, AutoMode{AutoOrder}, NoMetadata}

DimensionalData.AnonDim — Type

AnonDim <: Dimension

AnonDim()

Anonymous dimension. Used when extra dimensions are created, such as during transpose of a vector.

DimensionalData.@dim — Macro

@dim typ [supertype=Dimension] [name::String=string(typ)]

Macro to easily define new dimensions. The supertype will be inserted into the type of the dim. The default is simply YourDim <: Dimension. Making a Dimesion inherit from XDim, YDim, ZDim or TimeDim will affect automatic plot layout and other methods that dispatch on these types. <: YDim are plotted on the Y axis, <: XDim on the X axis, etc.

Example:

@dim Lat YDim "latitude"
@dim Lon XDim "Longitude"

Selectors

DimensionalData.Selector — Type

Selector

Abstract supertype for all selectors.

Selectors are wrappers that indicate that passed values are not the array indices, but values to be selected from the dimension index, such as DateTime objects for a Ti dimension.

Selectors provided in DimensionalData are:

At

Between

Near

Where

Contains

DimensionalData.At — Type

At <: Selector

At(x, atol, rtol)
At(x; atol=nothing, rtol=nothing)

Selector that exactly matches the value on the passed-in dimensions, or throws an error. For ranges and arrays, every intermediate value must match an existing value - not just the end points.

x can be any value or Vector of values.

atol and rtol are passed to isapprox. For Number rtol will be set to Base.rtoldefault, otherwise nothing, and wont be used.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(5:7)))
A[X(At(20)), Y(At(6))]

# output

5

DimensionalData.Near — Type

Near <: Selector

Near(x)

Selector that selects the nearest index to x.

With Points this is simply the index values nearest to the x, however with Intervals it is the interval center nearest to x. This will be offset from the index value for Start and End loci.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(5:7)))
A[X(Near(23)), Y(Near(5.1))]

# output

4

DimensionalData.Between — Type

Between <: Selector

Between(a, b)

Selector that retreive all indices located between 2 values, evaluated with >= for the lower value, and < for the upper value. This means the same value will not be counted twice in 2 adjacent Between selections.

For Intervals the whole interval must be lie between the values. For Points the points must fall between the values. Different Sampling types may give different results with the same input - this is the intended behaviour.

Between for Irregular intervals is a little complicated. The interval is the distance between a value and the next (for Start locus) or previous (for End locus) value.

For Center, we take the mid point between two index values as the start and end of each interval. This may or may not make sense for the values in your indes, so use Between with Irregular Intervals(Center()) with caution.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(5:7)))
A[X(Between(15, 25)), Y(Between(4, 6.5))]

# output

1×2 DimArray{Int64,2} with dimensions:
  X: 20:10:20 (Sampled - Ordered Regular Points)
  Y: 5:6 (Sampled - Ordered Regular Points)
 4  5

DimensionalData.Contains — Type

Contains <: Selector

Contains(x)

Selector that selects the interval the value is contained by. If the interval is not present in the index, an error will be thrown.

Can only be used for Intervals or Categorical.

Example

using DimensionalData

dims_ = X(10:10:20; mode=Sampled(sampling=Intervals())),
        Y(5:7; mode=Sampled(sampling=Intervals()))
A = DimArray([1 2 3; 4 5 6], dims_)
A[X(Contains(8)), Y(Contains(6.8))]

# output

3

DimensionalData.Where — Type

Where <: Selector

Where(f::Function)

Selector that filters a dimension by any function that accepts a single value from the index and returns a Bool.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(19:21)))
A[X(Where(x -> x > 15)), Y(Where(x -> x in (19, 21)))]

# output

1×2 DimArray{Int64,2} with dimensions:
  X: Int64[20] (Sampled - Ordered Regular Points)
  Y: Int64[19, 21] (Sampled - Ordered Regular Points)
 4  6

Name

DimensionalData.AbstractName — Type

AbstractName

Abstract supertype for name wrappers.

DimensionalData.Name — Type

Name <: AbstractName

Name(name::Union{Symbol,Name) => Name
Name(name::NoName) => NoName

Name wrapper. This lets arrays keep symbol names when the array wrapp neeeds to be `isbits, like for use on GPUs. It makes the name a property of the type. It's not necessary to use in normal use, a symbol is probably easier.

DimensionalData.NoName — Type

NoName <: AbstractName

NoName()

NoName specifies an array is not named, and is the default name value for all DimArrays. It can be used in set to remove the array name:

A = set(A, NoName())

Metadata

DimensionalData.AbstractMetadata — Type

AbstractMetadata{X,T}

Abstract supertype for all metadata wrappers.

Metadata wrappers allow tracking the contents and origin of metadata. This can facilitate conversion between metadata types (for saving a file to a differenet format) or simply saving data back to the same file type with identical metadata.

Using a wrapper instead of Dict or NamedTuple also lets us pass metadata objects to set without ambiguity about where to put them.

DimensionalData.Metadata — Type

Metadata <: AbstractMetadata

Metadata{X}(val::Union{Dict,NamedTuple})
Metadata{X}(pairs::Pair...) => Metadata{Dict}
Metadata{X}(; kw...) => Metadata{NamedTuple}

General Metadata object. The X type parameter categorises the metadata for method dispatch, if required.

DimensionalData.NoMetadata — Type

NoMetadata <: AbstractMetadata

NoMetadata()

Indicates an object has no metadata. Can be used in set to remove any existing metadata.

Modes

DimensionalData.Mode — Type

Mode

Abstract supertype for all Dimension modes. Defines or modifies dimension behaviour.

DimensionalData.IndexMode — Type

IndexMode <: Mode

Types defining the behaviour of a dimension index, how it is plotted and how Selectors like Between work.

An IndexMode may be a simple type like NoIndex indicating that the index is just the underlying array axis. It could also be a Categorical index indicating the index is ordered or unordered categories, or a Sampled index indicating sampling along some transect.

DimensionalData.Aligned — Type

Aligned <: Mode

Abstract supertype for IndexModes where the index is aligned with the array axes. This is by far the most common case.

DimensionalData.AbstractSampled — Type

AbstractSampled <: Aligned

Abstract supertype for IndexModes where the index is aligned with the array, and is independent of other dimensions. Sampled is provided by this package, Projected in GeoData.jl also extends AbstractSampled, adding crs projections.

AbstractSampled must have order, span and sampling fields, or a rebuild method that accpts them as keyword arguments.

DimensionalData.Sampled — Type

Sampled <: AbstractSampled

Sampled(order::Order, span::Span, sampling::Sampling)
Sampled(; order=AutoOrder(), span=AutoSpan(), sampling=Points())

A concrete implementation of the IndexMode AbstractSampled. It can be used to represent Points or Intervals.

It is capable of representing gridded data from a wide range of sources, allowing correct bounds and Selectors for points or intervals of regular, irregular, forward and reverse indexes.

The Sampled mode is assigned for all indexes of AbstractRange not assigned to Categorical.

Fields

order indicating array and index order (in Order), detected from the range order.

span indicates the size of intervals or distance between points, and will be set to Regular for AbstractRange and Irregular for AbstractArray, unless assigned manually.

sampling is assigned to Points, unless set to Intervals manually. Using Intervals will change the behaviour of bounds and Selectorss to take account for the full size of the interval, rather than the point alone.

Example

Create an array with [Interval] sampling.

using DimensionalData

dims_ = (X(100:-10:10; mode=Sampled(sampling=Intervals())),
         Y([1, 4, 7, 10]; mode=Sampled(span=Regular(2), sampling=Intervals())))
A = DimArray(rand(10, 4), dims_)
map(mode, dims(A))

# output

(Sampled{Ordered{ReverseIndex, ForwardArray, ForwardRelation}, Regular{Int64}, Intervals{Center}}(Ordered{ReverseIndex, ForwardArray, ForwardRelation}(ReverseIndex(), ForwardArray(), ForwardRelation()), Regular{Int64}(-10), Intervals{Center}(Center())), Sampled{Ordered{ForwardIndex, ForwardArray, ForwardRelation}, Regular{Int64}, Intervals{Center}}(Ordered{ForwardIndex, ForwardArray, ForwardRelation}(ForwardIndex(), ForwardArray(), ForwardRelation()), Regular{Int64}(2), Intervals{Center}(Center())))

DimensionalData.AbstractCategorical — Type

AbstractCategorical <: Aligned

IndexModes for dimensions where the values are categories.

Categorical is the provided concrete implementation.

AbstractCategorical must have an order field or a rebuild method with an order keyword argument.

DimensionalData.Categorical — Type

Categorical <: AbstractCategorical

Categorical(o::Order)
Categorical(; order=Unordered())

An IndexMode where the values are categories.

This will be automatically assigned if the index contains AbstractString, Symbol or Char. Otherwise it can be assigned manually.

Order will not be determined automatically for Categorical, it instead defaults to [Unordered].

Fields

order: Order indicating array and index order.

Example

Create an array with [Interval] sampling.

using DimensionalData

dims_ = X(["one", "two", "thee"]), Y([:a, :b, :c, :d])
A = DimArray(rand(3, 4), dims_)
map(mode, dims(A))

# output

(Categorical{Unordered{ForwardRelation}}(Unordered{ForwardRelation}(ForwardRelation())), Categorical{Unordered{ForwardRelation}}(Unordered{ForwardRelation}(ForwardRelation())))

DimensionalData.Unaligned — Type

Unaligned <: IndexMode

Abstract supertype for IndexMode where the Dimension index is not aligned to the grid.

Indexing with an Unaligned dimension with Selectors must provide all other Unaligned dimensions.

DimensionalData.Transformed — Type

Transformed <: Unaligned

Transformed(f, dim::Dimension)

IndexMode that uses an affine transformation to convert dimensions from dims(mode) to dims(array). This can be useful when the dimensions are e.g. rotated from a more commonly used axis.

Any function can be used to do the transformation, but transformations from CoordinateTransformations.jl may be useful.

Fields

f: transformation function

dims: a tuple containing dimenension types or symbols matching the order needed by the transform function.

Example

using DimensionalData, CoordinateTransformations

m = LinearMap([0.5 0.0; 0.0 0.5])
A = [1 2  3  4
     5 6  7  8
     9 10 11 12];
dimz = Dim{:t1}(mode=Transformed(m, X)),
              Dim{:t2}(mode=Transformed(m, Y))
da = DimArray(A, dimz)

da[X(At(6)), Y(At(2))]

# output

9

DimensionalData.NoIndex — Type

NoIndex <: Mode

NoIndex()

An IndexMode that is identical to the array axis.

Example

Defining a DimArray without passing an index to the dimension, the IndexMode will be NoIndex:

using DimensionalData

A = DimArray(rand(3, 3), (X, Y))
map(mode, dims(A))

# output

(NoIndex(), NoIndex())

Is identical to:

A = DimArray(rand(3, 3), (X(; mode=NoIndex()), Y(; mode=NoIndex())))
map(mode, dims(A))

# output

(NoIndex(), NoIndex())

DimensionalData.AutoMode — Type

AutoMode <: Mode

AutoMode()

Automatic IndexMode, the default mode. It will be converted automatically to another IndexMode when it is possible to detect it from the index.

Order of arrays and indices:

DimensionalData.ModeComponent — Type

ModeComponent

Abstract supertype of all component objects of Mode.

DimensionalData.Order — Type

Order <: ModeComponent

Traits for the order of the array, index and the relation between them.

DimensionalData.Unordered — Type

Unordered <: Order 

Unordered(relation=ForwardRelation())

Trait indicating that the array or dimension has no order. This means the index cannot be searched with searchsortedfirst, or similar methods, and that plotting order does not matter.

It still has a relation between the array axis and the dimension index.

DimensionalData.Ordered — Type

Ordered <: Order 

Ordered(index, array, relation)
Ordered(; index=ForwardIndex(), array=ForwardArray(), relation=ForwardRelation())

Container object for dimension and array ordering.

Fields

Each can have a value of ForwardX or ReverseX.

index: The order of the dimension index

array: The order of array axis, in terms of how you would want to plot it

relation: The relation between the index and the array.

All combinations of forward and reverse order for data and index seem to occurr in real datasets, as strange as that seems. We cover these possibilities by specifying the order of both explicitly, and the direction of the relationship between them.

Knowing the order of indices is important for using methods like searchsortedfirst() to find indices in sorted lists of values. Knowing the order of the data is then required to map to the actual indices. It's also used to plot the data later.

The default is Ordered(ForwardIndex(), ForwardArray(), ForwardRelation())

DimensionalData.AutoOrder — Type

AutoOrder <: Order

AutoOrder()

Order will be found automatically where possible.

This will fail for all dim eltypes without isless methods.

DimensionalData.UnknownOrder — Type

UnknownOrder <: Order

UnknownOrder()

Specifies that Order is not known and can't be determined.

DimensionalData.SubOrder — Type

SubOrder <: Order

Abstract supertype for sub-components of Order types

DimensionalData.IndexOrder — Type

IndexOrder <: SubOrder

Abstract supertype for dim index order

DimensionalData.ForwardIndex — Type

ForwardIndex <: IndexOrder

ForwardIndex()

Indicates that the dimension index is in the normal forward order.

DimensionalData.ReverseIndex — Type

ReverseIndex <: IndexOrder

ReverseIndex()

Indicates that the dimension index is in reverse order.

DimensionalData.ArrayOrder — Type

ArrayOrder <: IndexOrder

Abstract supertype for array ordering

DimensionalData.ForwardArray — Type

ForwardArray <: ArrayOrder

ForwardArray()

Indicates that the array axis is in the normal forward order.

DimensionalData.ReverseArray — Type

ReverseArray <: ArrayOrder

ReverseArray()

Indicates that the array axis is in reverse order.

It will be plotted backwards.

DimensionalData.Relation — Type

Relation <: SubOrder

Abstract supertype for index/array relationship

DimensionalData.ForwardRelation — Type

ForwardRelation <: Relation

ForwardRelation()

Indicates that the relationship between the index and the array is in the normal forward direction.

DimensionalData.ReverseRelation — Type

ReverseRelation <: Relation

ReverseRelation()

Indicates that the relationship between the index and the array is reversed.

Index modes for Intervals

Span

DimensionalData.Span — Type

Span <: ModeComponent

Defines the type of span used in a Sampling index. These are Regular or Irregular.

DimensionalData.Regular — Type

Regular <: Span

Regular(step=AutoStep())

Intervalss have regular size. This is passed to the constructor, although these are normally build automatically.

DimensionalData.Irregular — Type

Irregular <: Span

Irregular(bounds::Tuple)
Irregular(lowerbound, upperbound)

Irregular have irrigular size. To enable bounds tracking and accuract selectors, the starting bounds must be provided as a 2 tuple, or 2 arguments.

DimensionalData.Explicit — Type

Explicit(bounds::AbstractMatix)

Explicit span is explicitly listed for every interval. This uses a matrix where with length 2 columns for each index value - holding the lower and upper bounds for that specific index.

DimensionalData.AutoSpan — Type

AutoSpan <: Span

AutoSpan()

Span will be guessed and replaced by a constructor.

Sampling

DimensionalData.Sampling — Type

Sampling <: ModeComponent

Indicates the sampling method used by the index: Points or Intervals.

DimensionalData.Points — Type

Points <: Sampling

Points()

Sampling mode where single samples at exact points.

These are always plotted at the center of array cells.

DimensionalData.Intervals — Type

Intervals <: Sampling

Intervals(locus::Locus)

Sampling mode where samples are the mean (or similar) value over an interval.

Intervals require a Locus of Start, Center or End to define where in the interval the index values refer to.

Loci

Sampling positions for Intervals

DimensionalData.Locus — Type

Locus <: ModeComponent

Abstract supertype of types that indicate the position of index values in cells.

These allow for values array cells to align with the Start, Center, or End of values in the dimension index.

This means they can be plotted with correct axis markers, and allows automatic converrsions to between formats with different standards (such as NetCDF and GeoTiff).

Locii are often Start for time series, but often Center for spatial data.

These are reflected in the default values: Ti dimensions with Sampled index mode will default to Start Locii. All others default to Center.

DimensionalData.Center — Type

Center <: Locus

Center()

Indicates a dimension value is for the center of its corresponding array cell, in the direction of the dimension index order.

DimensionalData.Start — Type

Start <: Locus

Start()

Indicates a dimension value is for the start of its corresponding array cell, in the direction of the dimension index order.

DimensionalData.End — Type

End <: Locus

End()

Indicates a dimension value is for the end of its corresponding array cell, in the direction of the dimension index order.

DimensionalData.AutoLocus — Type

AutoLocus <: Locus

AutoLocus()

Indicates a dimension where the index position is not yet known. This will be filled with a default on object construction.

Tables.jl interface

DimensionalData.AbstractDimTable — Type

AbstractDimTable <: Tables.AbstractColumns

Abstract supertype for dim tables

DimensionalData.DimTable — Type

DimTable <: AbstractDimTable

DimTable(A::AbstractDimArray)

Construct a Tables.jl compatible object out of an AbstractDimArray.

This table will have a column for the array data and columns for each Dimension index, as a [DimColumn]. These are lazy, and generated as required.

Column names are converted from the dimension types using DimensionalData.dim2key. This means type Ti becomes the column name :Ti, and Dim{:custom} becomes :custom.

To get dimension columns, you can index with Dimension (X()) or Dimension type (X) as well as the regular Int or Symbol.

DimensionalData.DimColumn — Type

DimColumn{T,D<:Dimension} <: AbstractVector{T}

DimColumn(dim::Dimension, dims::Tuple{Vararg{<:DimTuple}})
DimColumn(dim::DimColumn, length::Int, dimstride::Int)

A table column based on a Dimension and it's relationship with other Dimensions in dims.

length is the product of all dim lengths (usually the length of the corresponding array data), while stride is the product of the preceding dimension lengths, which may or may not be the real stride of the corresponding array depending on the data type. For A isa Array, the dimstride will match the stride.

When the second argument is a Tuple of Dimension, the length and dimstride fields are calculated from the dimensions, relative to the column dimension dim.

This object will be returned as a column of DimTable.

Methods

Getting basic info

These useful functions for obtaining information from your dimensional data:

DimensionalData.dims — Function

dims(x) => Tuple{Vararg{<:Dimension}}
dims(x, dims::Tuple) => Tuple{Vararg{<:Dimension}}
dims(x, dim) => Dimension

Return a tuple of Dimensions for an object, in the order that matches the axes or columns etc. of the underlying data.

dims can be Dimension, Dimension types, or Symbols for Dim{Symbol}.

The default is to return nothing.

DimensionalData.refdims — Function

refdims(x) => Tuple{Vararg{<:Dimension}}

Reference dimensions for an array that is a slice or view of another array with more dimensions.

slicedims(a, dims) returns a tuple containing the current new dimensions and the new reference dimensions. Refdims can be stored in a field or disgarded, as it is mostly to give context to plots. Ignoring refdims will simply leave some captions empty.

The default is to return an empty Tuple ().

DimensionalData.metadata — Function

metadata(dim::Dimension)
metadata(dims::Tuple{<:Dimension,Vararg})
metadata(A::AbstractDimArray, dims::Tuple)  => (Dim metadata)
metadata(A::AbstractDimArray) => (Array metadata)

Returns the metadata for an array or the specified dimension(s). dims can be a Symbol (with Dim{X}, a Dimension, a Dimension type, or a mixed tuple.

dims can be a Dimension, a dimension type, or a tuple of either.

DimensionalData.name — Function

name(x) => String
name(xs::NTuple{N,<:Dimension}) => NTuple{N,String}
name(A::AbstractDimArray, dims::NTuple{N}) => NTuple{N,String}

Get the name of an array or Dimension, or a tuple of of either.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.units — Function

units(x) => Union{Nothing,Any}
units(::NTuple{N}) => NTuple{N}
unit(A::AbstractDimArray, dims::NTuple{N}) => NTuple{N,String}

Get the units of an array or Dimension, or a tuple of of either.

Units do not have a set field, and may or may not be included in metadata. This method is to facilitate use in labels and plots when units are available, not a guarantee that they will be. If not available, nothing is returned.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.label — Function

label(x) => String
label(dims::NTuple{N,<:Dimension}) => NTuple{N,String}
label(A::AbstractDimArray, dims::NTuple{N,<:Dimension}) => NTuple{N,String}

Get a plot label for data or a dimension. This will include the name and units if they exist, and anything else that should be shown on a plot.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

Dimesion and mode properties:

DimensionalData.val — Function

val(x)
val(dims::Tuple) => Tuple

Return the contained value of a wrapper object.

dims can be Dimension, Dimension types, or Symbols for Dim{Symbol}.

Objects that don't define a val method are returned unaltered.

DimensionalData.index — Function

index(dim::Dimension{<:Val}) => Tuple
index(dim::Dimension{<:AbstractArray}) => AbstractArray
index(dims::NTuple{N}) => Tuple{Vararg{Union{AbstractArray,Tuple},N}}

Return the contained index of a Dimension.

Only valid when a Dimension contains an AbstractArray or a Val tuple like Val{(:a, :b)}(). The Val is unwrapped to return just the Tuple

dims can be a Dimension, or a tuple of Dimension.

DimensionalData.mode — Function

mode(dim:Dimension) => IndexMode
mode(dims::Tuple) => Tuple{Vararg{<:IndexMode,N}}
mode(A::AbstractDimArray, [dims::Tuple]) => Tuple

Returns the IndexMode of a dimension. This dictates properties of the dimension such as array axis and index order, and sampling properties.

dims can be a Dimension, a dimension type, or a tuple of either.

DimensionalData.bounds — Function

bounds(dim::Dimension) => Tuple{T,T}}
bounds(dims::Tuple{<:Dimension,Vararg}) => Tuple{Vararg{<:Tuple{T,T}}}
bounds(A::AbstractArray, [dims]) => Tuple{Vararg{Tuple{T,T},N}}
bounds(A::AbstractArray, dim) => Tuple{T,T}

Return the bounds of all dimensions of an object, of a specific dimension, or of a tuple of dimensions.

dims can be a Dimension, a dimension type, or a tuple of either.

DimensionalData.sampling — Function

sampling(dim:Dimension) => Sampling
sampling(dims::Tuple) => Tuple{Vararg{<:Sampling,N}}
sampling(A::AbstractDimArray, [dims::Tuple]) => Tuple

Return the Sampling for each dimension.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.locus — Function

locus(dim:Dimension) => Locus
locus(dims::Tuple) => Tuple{Vararg{<:Locus,N}}
locus(A::AbstractDimArray, [dims::Tuple]) => Tuple

Return the Locus for each dimension.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.span — Function

span(dim:Dimension) => Span
span(dims::Tuple) => Tuple{Vararg{<:Span,N}}
span(A::AbstractDimArray, [dims::Tuple]) => Tuple

Return the Span for each dimension.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.order — Function

order(dim:Dimension) => Order
order(dims::Tuple) => Tuple{Vararg{<:Order,N}}
order(A::AbstractDimArray, [dims::Tuple]) => Tuple

Return the Order for each dimension.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.indexorder — Function

indexorder(dim:Dimension) => Union{Forward,Reverse}
indexorder(dims::Tuple) => Tuple{Vararg{<:Union{Forward,Reverse},N}}
indexorder(A::AbstractDimArray, [dims::Tuple]) => Tuple

Return the Order (Forward or Reverse) of the dimension index, for each dimension.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.arrayorder — Function

arrayorder(dim:Dimension) => Union{Forward,Reverse}
arrayorder(dims::Tuple) => Tuple{Vararg{<:Union{Forward,Reverse},N}}
arrayorder(A::AbstractDimArray, [dims::Tuple]) => Tuple

Return the Order (Forward or Reverse) of the array, for each dimension.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.relation — Function

relation(dim:Dimension) => Union{Forward,Reverse}
relation(dims::Tuple) => Tuple{Vararg{<:Union{Forward,Reverse},N}}
relation(A::AbstractDimArray, [dims::Tuple]) => Tuple

Return the relation (Forward or Reverse) between the dimension index and the array axis, for each dimension.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

Dimension querying

DimensionalData.hasdim — Function

hasdim([f], x, lookup::Tuple) => NTUple{Bool}
hasdim([f], x, lookups...) => NTUple{Bool}
hasdim([f], x, lookup) => Bool

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

lookup: Tuple or single Dimension or dimension Type.

f: <: by default, but can be >: to match abstract types to concrete types.

Check if an object or tuple contains an Dimension, or a tuple of dimensions.

Example

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

julia> hasdim(A, X)
true

julia> hasdim(A, (Z, X, Y))
(true, true, true)

julia> hasdim(A, Ti)
false

DimensionalData.dimnum — Function

dimnum(x, lookup::Tuple) => NTuple{Int}
dimnum(x, lookup) => Int

Get the number(s) of Dimension(s) as ordered in the dimensions of an object.

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

lookup: Tuple, Array or single Dimension or dimension Type.

The return type will be a Tuple of Int or a single Int, depending on wether lookup is a Tuple or single Dimension.

Example

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

julia> dimnum(A, (Z, X, Y))
(3, 1, 2)

julia> dimnum(A, Y)
2

DimensionalData.otherdims — Function

otherdims(x, lookup) => Tuple{Vararg{<:Dimension,N}}

Arguments

x: any object with a dims method, a Tuple of Dimension.

lookup: Tuple or single Dimension or dimension Type.

f: <: by default, but can be >: to match abstract types to concrete types.

A tuple holding the unmatched dimensions is always returned.

Example

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

julia> otherdims(A, X)
(Y{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()), Z{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()))

julia> otherdims(A, (Y, Z))
(X{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()),)

julia> otherdims(A, Ti)
(X{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()), Y{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()), Z{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()))

DimensionalData.commondims — Function

commondims([f], x, lookup) => Tuple{Vararg{<:Dimension}}

This is basically dims(x, lookup) where the order of the original is kept, unlike dims where the lookup tuple determines the order

Also unlike dims,commondims always returns a Tuple, no matter the input. No errors are thrown if dims are absent from either x or lookup.

f is <: by default, but can be >: to sort abstract types by concrete types.

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

julia> commondims(A, X)
(X{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()),)

julia> commondims(A, (X, Z))
(X{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()), Z{Base.OneTo{Int64}, NoIndex, NoMetadata}(Base.OneTo(10), NoIndex(), NoMetadata()))

julia> commondims(A, Ti)
()

And some utility methods for transforming DimensionalData objects:

DimensionalData.set — Function

set(x, val)

Set the field matching the supertypes of values in xs and return a new object.

As DimensionalData is so strongly typed you do not need to specify what field to set - there is no ambiguity.

To set fields of dimensions you need to specify the dimension. This can be done using Dimension => x pairs, X = x keyword arguments, Dimension wrapped arguments, or a NamedTuple.

When dimensions or IndexModes are passed to set to replace the existing ones, fields that are not set will keep their original values.

Notes:

Changing the dimension index range will set the Sampled mode component Regular with a new step size, and set the dimension order.

Setting Order will not reverse the array or dimension to match. Use reverse and reorder to do this.

Examples

da = DimArray(rand(3, 4), (Dim{:custom}(10.0:010.0:30.0), Z(-20:010.0:10.0)))

# Set the array values
set(da, zeros(3, 4))

# Set the array name
set(da, "newname") # Swap dimension type # Using Pairs set(da, :Z => Ti, :custom => Z) set(da, :custom => X, Z => Y)
# Using keyword arguments
set(da, custom = X, Z = :a)

# Using Dimension wrappers
set(da, Dim{:custom}(X), Z(Dim{:a}))

# Set the dimension index

# To an `AbstractArray` set(da, Z => [:a, :b, :c, :d], :custom => Val((4, 5, 6)))

# To a `Val` tuple index (compile time indexing)
set(da, Z(Val((:a, :b, :c, :d))), custom = 4:6)

# Set dim modes
set(da, Z=NoIndex(), custom=Sampled())
set(da, :custom => Irregular(10, 12), Z => Regular(9.9))
set(da, (Z=NoIndex(), custom=Sampled()))
set(da, custom=Ordered(array=Reverse()), Z=Unordered())

DimensionalData.rebuild — Function

rebuild(x, args...)
rebuild(x; kw...)

Rebuild an object struct with updated field values.

This is an abstraction that alows inbuilt and custom types to be rebuilt functionally to update them, as most objects in DimensionalData are immutable.

x can be a AbstractDimArray, a Dimension, IndexMode or other custom types.

The arguments reuired are defined for the abstract type that has a rebuild method.

DimensionalData.modify — Function

modify(f, A::AbstractDimArray) => AbstractDimArray
modify(f, s::AbstractDimStack) => AbstractDimStack
modify(f, dim::Dimension) => Dimension
modify(f, x, lookupdim::Dimension) => typeof(x)

Modify the parent data, rebuilding the object wrapper without change. f must return a AbstractArray of the same size as the original.

This method is mostly useful as a way of swapping the parent array type of an object.

Example

If we have a previously-defined DimArray, we can copy it to an Nvidia GPU with:

A = DimArray(rand(100, 100), (X, Y))
modify(CuArray, A)

This also works for all the data layers in a DimStack.

DimensionalData.dimwise — Function

dimwise(f, A::AbstractDimArray{T,N}, B::AbstractDimArray{T2,M}) => AbstractDimArray{T3,N}

Dimension-wise application of function f to A and B.

Arguments

a: AbstractDimArray to broacast from, along dimensions not in b.

b: AbstractDimArray to broadcast from all dimensions. Dimensions must be a subset of a.

This is like broadcasting over every slice of A if it is sliced by the dimensions of B.

DimensionalData.dimwise! — Function

dimwise!(f, dest::AbstractDimArray{T1,N}, A::AbstractDimArray{T2,N}, B::AbstractDimArray) => dest

Dimension-wise application of function f.

Arguments

dest: AbstractDimArray to update

a: AbstractDimArray to broacast from, along dimensions not in b.

b: AbstractDimArray to broadcast from all dimensions. Dimensions must be a subset of a.

This is like broadcasting over every slice of A if it is sliced by the dimensions of B, and storing the value in dest.

DimensionalData.reorder — Function

reorder(::order, A::AbstractDimArray) => AbstractDimArray
reorder(A::AbstractDimArray, order::Union{Order,Dimension{<:Order},Tuple}) => AbstractDimArray
reorder(A::AbstractDimArray, order::Pair{<:Dimension,<:SubOrder}...) => AbstractDimArray

Reorder every dims index/array/relation to order, or reorder index for the the given dimension(s) to the Order they wrap.

Reorderind Relation will reverse the array, not the dimension index.

order can be an Order, a single Dimension or a Tuple of Dimension.

Base.reverse — Function

reverse(A; dims) => AbstractDimArray
reverse(dim::Dimension) => Dimension

Reverse the array order, and update the dim to match.

Base.fill — Function

Base.fill(x::T, dims::Vararg{Dimension,N}}) => DimArray{T,N}
Base.fill(x::T, dims::Tuple{Vararg{Dimension,N}}) => DimArray{T,N}

Create a DimArray from a fill value x and Dimensions.

A Dimension with an AbstractVector value will set the array axis

A Dimension holding an Integer will set the length of the Array axis, and set the dimension mode to NoIndex.

Base.map — Function

Base.map(f, s::AbstractDimStack)

Apply functrion f to each layer of the stack s, and rebuild it.

If f returns DimArrays the result will be another DimStack. Other values will be returned in a NamedTuple.

Base.copy! — Function

Base.copy!(dst::AbstractDimStack, src::AbstractDimStack, [keys=keys(dst)])

Copy all or a subset of layers from one stack to another.

Example

Copy just the :sea_surface_temp and :humidity layers from src to dst.

copy!(dst::AbstractDimStack, src::AbstractDimStack, keys=(:sea_surface_temp, :humidity))

Non-exported methods for developers

DimensionalData.dim2key — Function

dim2key(dim::Dimension) => Symbol
dim2key(dims::Type{<:Dimension}) => Symbol

Convert a dimension object to a simbol. X(), Y(), Ti() etc will be converted. to :X, :Y, :Ti, as with any other dims generated with the @dim macro.

All other Dim{S}() dimensions will generate Symbols S.

DimensionalData.key2dim — Function

key2dim(s::Symbol) => Dimension
key2dim(dims...) => Tuple{Dimension,Vararg}
key2dim(dims::Tuple) => Tuple{Dimension,Vararg}

Convert a symbol to a dimension object. :X, :Y, :Ti etc will be converted. to X(), Y(), Ti(), as with any other dims generated with the @dim macro.

All other Symbols S will generate Dim{S}() dimensions.

DimensionalData.dims2indices — Function

dims2indices(dim::Dimension, lookup) => NTuple{Union{Colon,AbstractArray,Int}}

Convert a Dimension or Selector lookup to indices of Int, AbstractArray or Colon.

DimensionalData.formatdims — Function

formatdims(A, dims) => Tuple{Vararg{<:Dimension,N}}

Format the passed-in dimension(s) dims to match the array A.

This means converting indexes of Tuple to LinRange, and running identify. Errors are also thrown if dims don't match the array dims or size.

If a IndexMode hasn't been specified, an mode is chosen based on the type and element type of the index:

DimensionalData.reducedims — Function

reducedims(x, dimstoreduce)

Replace the specified dimensions with an index of length 1. This is usually to match a new array size where an axis has been reduced with a method like mean or reduce to a length of 1, but the number of dimensions has not changed.

IndexMode traits are also updated to correspond to the change in cell step, sampling type and order.

DimensionalData.swapdims — Function

swapdims(x::T, newdims) => T
swapdims(dims::Tuple, newdims) => Tuple{Dimension}

Swap dimensions for the passed in dimensions, in the order passed.

Passing in the Dimension types rewraps the dimension index, keeping the index values and metadata, while constructed Dimension objectes replace the original dimension. nothing leaves the original dimension as-is.

Arguments

x: any object with a dims method or a Tuple of Dimension.

newdim: Tuple of Dimension or dimension Type.

Example

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

DimensionalData.slicedims — Function

slicedims(x, I) => Tuple{Tuple,Tuple}
slicedims(f, x, I) => Tuple{Tuple,Tuple}

Slice the dimensions to match the axis values of the new array.

All methods return a tuple conatining two tuples: the new dimensions, and the reference dimensions. The ref dimensions are no longer used in the new struct but are useful to give context to plots.

Called at the array level the returned tuple will also include the previous reference dims attached to the array.

Arguments

f: a function getindex, view or dotview. This will be used for slicing getindex is the default if f is not included.

x: An AbstractDimArray, Tuple of Dimension, or Dimension

I: A tuple of Integer, Colon or AbstractArray

DimensionalData.comparedims — Function

comparedims(A::AbstractDimArray...)
comparedims(A::Tuple...)
comparedims(a, b)

Check that dimensions or tuples of dimensions are the same, and return the first valid dimension. If AbstractDimArrays are passed as arguments their dimensions are compared.

Empty tuples and nothing dimension values are ignored, returning the Dimension value if it exists.

DimensionalData.sortdims — Function

sortdims([f], tosort, order) => Tuple

Sort dimensions tosort by order. Dimensions in order but missing from tosort are replaced with nothing.

tosort and order can be Tuples or Vectors or Dimension or dimension type. Abstract supertypes like TimeDim can be used in order.

f is <: by default, but can be >: to sort abstract types by concrete types.

DimensionalData.identify — Function

identify(indexmode, index)

Identify an IndexMode and its fields from the index content and the existing IndexMode.

These methods guess which mode is most appropriate, encoding information about the index in types that can later be used for dispatch. They also let you fill in part of the information you need to specify, and guess the rest.

An example of the usefulness of identifying index traits up-front is if we check that the index is ordered, we can be sure it remains ordered unless it is indexed with a Vector. This means we will always use the correct searchsorted method for it.

DimensionalData.basetypeof — Function

basetypeof(x) => Type

Get the "base" type of an object - the minimum required to define the object without it's fields. By default this is the full UnionAll for the type. But custom basetypeof methods can be defined for types with free type parameters.

In DimensionalData this is primariliy used for comparing Dimensions, where Dim{:x} is different from Dim{:y}.

DimensionalData.setdims — Function

setdims(A::AbstractArray, newdims) => AbstractArray
setdims(::Tuple, newdims) => Tuple{Vararg{<:Dimension,N}}

Replaces the first dim matching <: basetypeof(newdim) with newdim, and returns a new object or tuple with the dimension updated.

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

newdim: Tuple or single Dimension or dimension Type.

Example

using DimensionalData

A = DimArray(ones(10, 10), (X, Y(10:10:100)))
B = setdims(A, Y('a':'j'))
val(dims(B, Y))

# output

'a':1:'j'

DimensionalData.flip — Function

fliparray(Order, A, dims) => AbstractDimArray
fliparray(dim::Dimension) => Dimension

Flip the array order without changing any data.

DimensionalData.dimsmatch — Function

dimsmatch([f], dim, lookup) => Bool
dimsmatch([f], dims::Tuple, lookups::Tuple) => Bool

Compare 2 dimensions or Tuple of Dimension are of the same base type, or are at least rotations/transformations of the same type.

f is <: by default, but can be >: to match abstract types to concrete types.

DimensionalData.dimstride — Function

dimstride(x, dim)

Will get the stride of the dimension relative to the other dimensions.

This may or may not be equal to the stride of the related array, although it will be for Array.

Arguments

x is any object with a dims method, or a Tuple of Dimension.

dim is a Dimension, Dimension type, or and Int. Using an Int is not type-stable.