API

To use the functionality of DimensionalData in your module, dispatch on AbstractDimArray and AbstractDimension.

Arrays:

DimensionalData.AbstractDimArray — Type

Supertype for all "dim" arrays.

These arrays return a Tuple of Dimension from a dims method, and can be rebuilt using rebuild.

parent must return the source array.

They should have metadata, name and refdims methods, although these are optional.

A rebuild method for AbstractDimArray must accept data, dims, refdims, name, metadata arguments.

DimensionalData.DimArray — Type

DimArray(data, dims, refdims, name)

The main subtype of AbstractDimArray. Maintains and updates its dimensions through transformations and moves dimensions to refdims after reducing operations (like e.g. mean).

Core types

Dimensions:

DimensionalData.Dimension — Type

Supertype of all dimension types.

Example concrete implementations are X, Y, Z, Ti (Time), and the custom [Dim]@ref) dimension.

Dimensions label the axes of an AbstractDimesnionalArray, or other dimensional objects, and are used to index into the array.

They may also provide an alternate index to lookup for each array axis. This may be any AbstractVector matching the array axis length, or a Val holding a tuple for compile-time index lookups.

Dimensions also have mode and metadata fields.

mode gives more details about the dimension, such as that it is Categorical or Sampled as Points or Intervals along some transect. DimensionalData will attempt to guess the mode from the passed-in index value.

metadata can hold any metadata object adding more information about the array axis - useful for extending DimensionalData for specific contexts, like geospatial data in GeoData.jl. By default it is nothing.

Example:

using DimensionalData, Dates

x = X(2:2:10)
y = Y(['a', 'b', 'c'])
ti = Ti(DateTime(2021, 1):Month(1):DateTime(2021, 12))

A = DimArray(zeros(3, 5, 12), (y, x, ti))

# output

DimArray with dimensions:
 Y: Char[a, b, c] (Categorical: Unordered)
 X: 2:2:10 (Sampled: Ordered Regular Points)
 Time (type Ti): DateTime("2021-01-01T00:00:00"):Month(1):DateTime("2021-12-01T00:00:00") (Sampled: Ordered Regular Points)
and data: 3×5×12 Array{Float64,3}
[:, :, 1]
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
[and 11 more slices...]

For simplicity, the same Dimension types are also used as wrappers in getindex, like:

x = A[X(2), Y(3)]

# output

DimArray with dimensions:
 Time (type Ti): DateTime("2021-01-01T00:00:00"):Month(1):DateTime("2021-12-01T00:00:00") (Sampled: Ordered Regular Points)
and referenced dimensions:
 Y: c (Categorical: Unordered)
 X: 4 (Sampled: Ordered Regular Points)
and data: 12-element Array{Float64,1}
[0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]

A Dimension can also wrap Selector.

x = A[X(Between(3, 4)), Y(At('b'))]

# output

DimArray with dimensions:
 X: 4:2:4 (Sampled: Ordered Regular Points)
 Time (type Ti): DateTime("2021-01-01T00:00:00"):Month(1):DateTime("2021-12-01T00:00:00") (Sampled: Ordered Regular Points)
and referenced dimensions:
 Y: b (Categorical: Unordered)
and data: 1×12 Array{Float64,2}
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

Dimension objects may have mode and metadata fields to track additional information about the data and the index, and their relationship.

DimensionalData.DependentDim — Type

Supertype for Dependent dimensions. These will plot on the Y axis.

DimensionalData.IndependentDim — Type

Supertype for independent dimensions. Thise will plot on the X axis.

DimensionalData.XDim — Type

Supertype for all X dimensions.

DimensionalData.YDim — Type

Supertype for all Y dimensions.

DimensionalData.ZDim — Type

Supertype for all Z dimensions.

DimensionalData.TimeDim — Type

Supertype for all time dimensions.

In a TimeDime with Interval sampling the locus will automatically be set to Start(). Dates and times generally refer to the start of a month, hour, second etc., not the central point as is more common with spatial data. `

DimensionalData.X — Type

X(val=:; mode=AutoMode(), metadata=nothing)

X Dimension. X <: XDim <: IndependentDim

Example:

xdim = X(2:2:10)
# Or
val = A[X(1)]
# Or
mean(A; dims=X)

DimensionalData.Y — Type

Y(val=:; mode=AutoMode(), metadata=nothing)

Y Dimension. Y <: YDim <: DependentDim

Example:

ydim = Y(['a', 'b', 'c'])
# Or
val = A[Y(1)]
# Or
mean(A; dims=Y)

DimensionalData.Z — Type

Z(val=:; mode=AutoMode(), metadata=nothing)

Z Dimension. Z <: ZDim <: Dimension

Example:

zdim = Z(10:10:100)
# Or
val = A[Z(1)]
# Or
mean(A; dims=Z)

DimensionalData.Ti — Type

Ti(val=:; mode=AutoMode(), metadata=nothing)

Time Dimension. Ti <: TimeDim <: IndependentDim

Time is already used by Dates, and T is a common type parameter, We use Ti to avoid clashes.

Example:

timedim = Ti(DateTime(2021, 1):Month(1):DateTime(2021, 12))
# Or
val = A[Ti(1)]
# Or
mean(A; dims=Ti)

DimensionalData.ParametricDimension — Type

Supertype for Dimensions with user-set type paremeters

DimensionalData.Dim — Type

Dim{:X}()
Dim{:X}(val=:; mode=AutoMode(), metadata=nothing)
Dim{:X}(val, mode, metadata=nothing)

A generic dimension. For use when custom dims are required when loading data from a file. The sintax is ugly and verbose to use for indexing, ie Dim{:lat}(1:9) rather than Lat(1:9). This is the main reason they are not the only type of dimension availabile.

using DimensionalData

dim = Dim{:custom}(['a', 'b', 'c'])

# output

dimension Dim{:custom} (type Dim):
val: Char[a, b, c]
mode: AutoMode
metadata: nothing
type: Dim{:custom,Array{Char,1},AutoMode{AutoOrder},Nothing}

DimensionalData.AnonDim — Type

AnonDim()

Anonymous dimension. Used when extra dimensions are created, such as during transpose of a vector.

DimensionalData.@dim — Macro

@dim typ [supertype=Dimension] [name=string(typ)] [shortname=string(typ)]

Macro to easily define new dimensions. The supertype will be inserted into the type of the dim. The default is simply YourDim <: Dimension. Making a Dimesion inherit from XDim, YDim, ZDim or TimeDim will affect automatic plot layout and other methods that dispatch on these types. <: YDim are plotted on the Y axis, <: XDim on the X axis, etc.

Example:

@dim Lat "Lattitude" "lat"
@dim Lon XDim "Longitude"

Selectors

DimensionalData.Selector — Type

Selectors are wrappers that indicate that passed values are not the array indices, but values to be selected from the dimension index, such as DateTime objects for a Ti dimension.

Selectors provided in DimensionalData are:

At

Between

Near

Where

Contains

DimensionalData.At — Type

At(x, atol, rtol)
At(x; atol=nothing, rtol=nothing)

Selector that exactly matches the value on the passed-in dimensions, or throws an error. For ranges and arrays, every intermediate value must match an existing value - not just the end points.

x can be any value or Vector of values.

atol and rtol are passed to isapprox. For Number rtol will be set to Base.rtoldefault, otherwise nothing, and wont be used.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(5:7)))
A[X(At(20)), Y(At(6))]

# output

5

DimensionalData.Near — Type

Near(x)

Selector that selects the nearest index to x.

With Points this is simply the index values nearest to the x, however with Intervals it is the interval center nearest to x. This will be offset from the index value for Start and End loci.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(5:7)))
A[X(Near(23)), Y(Near(5.1))]

# output

4

DimensionalData.Between — Type

Between(a, b)

Selector that retreive all indices located between 2 values, evaluated with >= for the lower value, and < for the upper value. This means the same value will not be counted twice in 2 Between selections.

For Intervals the whole interval must be lie between the values. For Points the points must fall between the values. Different Sampling types may give different results with the same input - this is the intended behaviour.

Between for Irregular intervals is a little complicated. The interval is the distance between a value and the next (for Start locus) or previous (for End locus) value.

For Center, we take the mid point between two index values as the start and end of each interval. This may or may not make sense for the values in your indes, so use Between with Irregular Intervals(Center()) with caution.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(5:7)))
A[X(Between(15, 25)), Y(Between(4, 6.5))]

# output

DimArray with dimensions:
 X: 20:10:20 (Sampled: Ordered Regular Points)
 Y: 5:6 (Sampled: Ordered Regular Points)
and data: 1×2 Array{Int64,2}
 4  5

DimensionalData.Contains — Type

Contains(x)

Selector that selects the interval the value is contained by. If the interval is not present in the index, an error will be thrown.

Can only be used for Intervals or Categorical.

Example

using DimensionalData

dims_ = X(10:10:20; mode=Sampled(sampling=Intervals())),
        Y(5:7; mode=Sampled(sampling=Intervals()))
A = DimArray([1 2 3; 4 5 6], dims_)
A[X(Contains(8)), Y(Contains(6.8))]

# output

3

DimensionalData.Where — Type

Where(f::Function)

Selector that filters a dimension by any function that accepts a single value from the index and returns a Bool.

Example

using DimensionalData

A = DimArray([1 2 3; 4 5 6], (X(10:10:20), Y(19:21)))
A[X(Where(x -> x > 15)), Y(Where(x -> x in (19, 21)))]

# output

DimArray with dimensions:
 X: Int64[20] (Sampled: Ordered Regular Points)
 Y: Int64[19, 21] (Sampled: Ordered Regular Points)
and data: 1×2 Array{Int64,2}
 4  6

Index Modes

DimensionalData.IndexMode — Type

Types defining the behaviour of a dimension, how they are plotted and how Selectors like Between work on them.

An IndexMode may be a simple type like NoIndex indicating that the index is just the underlying array axis. It could also be a Categorical index indicating the index is ordered or unordered categories, or a Sampled index indicating sampling along some transect.

DimensionalData.Aligned — Type

Supertype for IndexModes where the index is aligned with the array axes. This is by far the most common case.

DimensionalData.AbstractSampled — Type

Abstract supertype for IndexModes where the index is aligned with the array, and is independent of other dimensions. Sampled is provided by this package, Projected in GeoData.jl also extends AbstractSampled, adding crs projections.

A rebuild method for AbstractSampled must accept order, span and sampling, arguments.

DimensionalData.Sampled — Type

Sampled(order::Order, span::Span, sampling::Sampling)
Sampled(; order=AutoOrder(), span=AutoSpan(), sampling=Points())

A concrete implementation of the IndexMode AbstractSampled. It can be used to represent Points or Intervals.

It is capable of representing gridded data from a wide range of sources, allowing correct bounds and Selectors for points or intervals of regular, irregular, forward and reverse indexes.

The Sampled mode is assigned for all indexes of AbstractRange not assigned to Categorical.

Fields

order indicating array and index order (in Order), detected from the range order.

span indicates the size of intervals or distance between points, and will be set to Regular for AbstractRange and Irregular for AbstractArray, unless assigned manually.

sampling is assigned to Points, unless set to Intervals manually. Using Intervals will change the behaviour of bounds and Selectorss to take account for the full size of the interval, rather than the point alone.

Example

Create an array with [Interval] sampling.

using DimensionalData

dims_ = (X(100:-10:10; mode=Sampled(sampling=Intervals())),
         Y([1, 4, 7, 10]; mode=Sampled(span=Regular(2), sampling=Intervals())))
A = DimArray(rand(10, 4), dims_)
map(mode, dims(A))

# output

(Sampled: Ordered Regular Intervals, Sampled: Ordered Regular Intervals)

DimensionalData.AbstractCategorical — Type

IndexModes for dimensions where the values are categories.

Categorical is the provided concrete implementation.

A rebuild method for AbstractCategorical must accept the order argumen.

DimensionalData.Categorical — Type

Categorical(o::Order)
Categorical(; order=Unordered())

An IndexMode where the values are categories.

This will be automatically assigned if the index contains AbstractString, Symbol or Char. Otherwise it can be assigned manually.

Order will not be determined automatically for Categorical, it instead defaults to [Unordered].

Fields

order: Order indicating array and index order.

Example

Create an array with [Interval] sampling.

using DimensionalData

dims_ = X(["one", "two", "thee"]), Y([:a, :b, :c, :d])
A = DimArray(rand(3, 4), dims_)
map(mode, dims(A))

# output

(Categorical: Unordered, Categorical: Unordered)

DimensionalData.Unaligned — Type

Supertype for IndexMode where the Dimension index is not aligned to the grid.

Indexing with an Unaligned dimension with Selectors must provide all other Unaligned dimensions.

DimensionalData.Transformed — Type

Transformed(f, dim::Dimension)

IndexMode that uses an affine transformation to convert dimensions from dims(mode) to dims(array). This can be useful when the dimensions are e.g. rotated from a more commonly used axis.

Any function can be used to do the transformation, but transformations from CoordinateTransformations.jl may be useful.

Fields

f: transformation function

dims: a tuple containing dimenension types or symbols matching the order needed by the transform function.

Example

using DimensionalData, CoordinateTransformations

m = LinearMap([0.5 0.0; 0.0 0.5])
A = [1 2  3  4
     5 6  7  8
     9 10 11 12];
dimz = Dim{:t1}(mode=Transformed(m, X)),
              Dim{:t2}(mode=Transformed(m, Y))
da = DimArray(A, dimz)

da[X(At(6)), Y(At(2))]

# output

9

DimensionalData.NoIndex — Type

NoIndex()

An IndexMode that is identical to the array axis.

Example

Defining a DimArray without passing an index to the dimension, the IndexMode will be NoIndex:

using DimensionalData

A = DimArray(rand(3, 3), (X, Y))
map(mode, dims(A))

# output

(NoIndex, NoIndex)

Is identical to:

A = DimArray(rand(3, 3), (X(; mode=NoIndex()), Y(; mode=NoIndex())))
map(mode, dims(A))

# output

(NoIndex, NoIndex)

DimensionalData.AutoMode — Type

AutoMode()

Automatic IndexMode, the default mode. It will be converted automatically to another IndexMode when it is possible to detect it from the index.

Order of arrays and indices:

DimensionalData.Order — Type

Traits for the order of the array, index and the relation between them.

DimensionalData.Unordered — Type

Unordered(relation=Forward())

Trait indicating that the array or dimension has no order. This means the index cannot be searched with searchsortedfirst, or similar methods, and that plotting order does not matter.

It still has a relation between the array axis and the dimension index.

DimensionalData.Ordered — Type

Ordered(index, array, relation)
Ordered(; index=Forward(), array=Forward(), relation=Forward())

Container object for dimension and array ordering.

Fields

Each can have a value of Forward or Reverse.

index: The order of the dimension index

array: The order of array axis, in terms of how you would want to plot it

relation: The relation between the index and the array.

All combinations of forward and reverse order for data and index seem to occurr in real datasets, as strange as that seems. We cover these possibilities by specifying the order of both explicitly, and the direction of the relationship between them.

Knowing the order of indices is important for using methods like searchsortedfirst() to find indices in sorted lists of values. Knowing the order of the data is then required to map to the actual indices. It's also used to plot the data later.

The default is Ordered(Forward(), Forward(), Forward())

DimensionalData.AutoOrder — Type

AutoOrder()

Order will be found automatically where possible.

This will fail for all dim eltypes without isless methods.

DimensionalData.UnknownOrder — Type

UnknownOrder()

Order is not known and can't be determined.

DimensionalData.Forward — Type

Forward()

Indicates that the array axis, dimension index or the relation between them is in the normal forward order.

DimensionalData.Reverse — Type

Reverse()

Indicates that the array axis, dimension index or the relation between them is in reverse order.

Index modes for Intervals

Span

DimensionalData.Span — Type

Defines the type of span used in a Sampling index. These are Regular or Irregular.

DimensionalData.Regular — Type

Regular(step=AutoStep())

Intervalss have regular size. This is passed to the constructor, although these are normally build automatically.

DimensionalData.Irregular — Type

Irregular(bounds::Tuple)
Irregular(lowerbound, upperbound)

Irregular have irrigular size. To enable bounds tracking and accuract selectors, the starting bounds must be provided as a 2 tuple, or 2 arguments.

DimensionalData.AutoSpan — Type

AutoSpan()

Span will be guessed and replaced by a constructor.

Sampling

DimensionalData.Sampling — Type

Indicates the sampling method used by the index: Points or Intervals.

DimensionalData.Points — Type

Points()

Sampling mode where single samples at exact points.

These are always plotted at the center of array cells.

DimensionalData.Intervals — Type

Intervals(locus::Locus)

Sampling mode where samples are the mean (or similar) value over an interval.

Intervals require a Locus of Start, Center or End to define where in the interval the index values refer to.

Loci

Sampling positions for Intervals

DimensionalData.Locus — Type

Locii indicate the position of index values in cells.

These allow for values array cells to align with the Start, Center, or End of values in the dimension index.

This means they can be plotted with correct axis markers, and allows automatic converrsions to between formats with different standards (such as NetCDF and GeoTiff).

Locii are often Start for time series, but often Center for spatial data.

These are reflected in the default values: Ti dimensions with Sampled index mode will default to Start Locii. All others default to Center.

DimensionalData.Center — Type

Center()

Indicates a dimension value is for the center of its corresponding array cell, in the direction of the dimension index order.

DimensionalData.Start — Type

Start()

Indicates a dimension value is for the start of its corresponding array cell, in the direction of the dimension index order.

DimensionalData.End — Type

End()

Indicates a dimension value is for the end of its corresponding array cell, in the direction of the dimension index order.

DimensionalData.AutoLocus — Type

AutoLocus()

Indicates a dimension where the index position is not yet known. This will be filled with a default on object construction.

Methods

Getting basic info

These useful functions for obtaining information from your dimensional data:

DimensionalData.bounds — Function

bounds(dim::Dimension) => Tuple{T,T}}
bounds(dims::Tuple{<:Dimension,Vararg}) => Tuple{Vararg{<:Tuple{T,T}}}
bounds(A::AbstractArray, [dims]) => Tuple{Vararg{Tuple{T,T},N}}
bounds(A::AbstractArray, dim) => Tuple{T,T}

Return the bounds of all dimensions of an object, of a specific dimension, or of a tuple of dimensions.

dims can be a Dimension, a dimension type, or a tuple of either.

Missing docstring.

Missing docstring for data. Check Documenter's build log for details.

DimensionalData.dimnum — Function

dimnum(x, lookup::Tuple) => NTuple{Int}
dimnum(x, lookup) => Int

Get the number(s) of Dimension(s) as ordered in the dimensions of an object.

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

lookup: Tuple, Array or single Dimension or dimension Type.

The return type will be a Tuple of Int or a single Int, depending on wether lookup is a Tuple or single Dimension.

Example

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

DimensionalData.dims — Function

dims(x) => Tuple{Vararg{<:Dimension}}
dims(x, dims::Tuple) => Tuple{Vararg{<:Dimension}}
dims(x, dim) => Dimension

Return a tuple of Dimensions for an object, in the order that matches the axes or columns etc. of the underlying data.

dims can be Dimension, Dimension types, or Symbols for Dim{Symbol}.

The default is to return nothing.

DimensionalData.hasdim — Function

hasdim(x, lookup::Tuple) => NTUple{Bool}
hasdim(x, lookup) => Bool

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

lookup: Tuple or single Dimension or dimension Type.

Check if an object or tuple contains an Dimension, or a tuple of dimensions.

Example

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

DimensionalData.otherdims — Function

otherdims(x, lookup) => Tuple{Vararg{<:Dimension,N}}

Arguments

x: any object with a dims method, a Tuple of Dimension.

lookup: Tuple or single Dimension or dimension Type.

A tuple holding the unmatched dimensions is always returned.

Example

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

DimensionalData.commondims — Function

commondims(x, lookup) => Tuple{Vararg{<:Dimension}}

This is basically dims(x, lookup) where the order of the original is kept, unlike dims where the lookup tuple determines the order

Also unlike dims,commondims always returns a Tuple, no matter the input.

DimensionalData.label — Function

label(x) => String
label(dims::NTuple{N,<:Dimension}) => NTuple{N,String}
label(A::AbstractDimArray, dims::NTuple{N,<:Dimension}) => NTuple{N,String}

Get a plot label for data or a dimension. This will include the name and units if they exist, and anything else that should be shown on a plot.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.mode — Function

mode(dim:Dimension) => IndexMode 
mode(dims::Tuple) => Tuple{Vararg{<:IndexMode,N}}
mode(A::AbstractDimArray, [dims::Tuple]) => Tuple

Returns the IndexMode of a dimension. This dictates properties of the dimension such as array axis and index order, and sampling properties.

dims can be a Dimension, a dimension type, or a tuple of either.

DimensionalData.metadata — Function

metadata(dim::Dimension)
metadata(dims::Tuple{<:Dimension,Vararg})
metadata(A::AbstractDimArray, dims::Tuple)  => (Dim metadata)
metadata(A::AbstractDimArray) => (Array metadata)

Returns the metadata for an array or the specified dimension(s). dims can be a Symbol (with Dim{X}, a Dimension, a Dimension type, or a mixed tuple.

dims can be a Dimension, a dimension type, or a tuple of either.

DimensionalData.name — Function

name(x) => String
name(xs::NTuple{N,<:Dimension}) => NTuple{N,String}
name(A::AbstractDimArray, dims::NTuple{N}) => NTuple{N,String}

Get the name of an array or Dimension, or a tuple of of either.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.refdims — Function

refdims(x) => Tuple{Vararg{<:Dimension}}

Reference dimensions for an array that is a slice or view of another array with more dimensions.

slicedims(a, dims) returns a tuple containing the current new dimensions and the new reference dimensions. Refdims can be stored in a field or disgarded, as it is mostly to give context to plots. Ignoring refdims will simply leave some captions empty.

The default is to return an empty Tuple ().

DimensionalData.shortname — Function

shortname(x) => String
shortname(xs::NTuple{N}) => NTuple{N,String}
shortname(A::AbstractDimArray, dims::NTuple{N}) => NTuple{N,String}

Get the shortname of an array or Dimension, or a tuple of of either.

This may be a shorter version more suitable for small labels than name, but it may also be identical to name.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.units — Function

units(x) => Union{Nothing,Any}
units(::NTuple{N}) => NTuple{N}
unit(A::AbstractDimArray, dims::NTuple{N}) => NTuple{N,String}

Get the units of an array or Dimension, or a tuple of of either.

Units do not have a set field, and may or may not be included in metadata. This method is to facilitate use in labels and plots when units are available, not a guarantee that they will be. If not available, nothing is returned.

dims can be Dimensions, Dimension types, or Symbols for Dim{Symbol}.

DimensionalData.val — Function

val(x)
val(dims::Tuple) => Tuple
val(A::AbstractDimArray, dims::Tuple)  => Tuple

Return the contained value of a wrapper object.

dims can be Dimension, Dimension types, or Symbols for Dim{Symbol}.

Objects that don't define a val method are returned unaltered.

Missing docstring.

Missing docstring for basetypeof. Check Documenter's build log for details.

And some utility methods for transforming DimensionalData objects:

DimensionalData.rebuild — Function

rebuild(x, args...)
rebuild(x; kwargs...)

Rebuild an object struct with updated field values.

This is an abstraction that alows inbuilt and custom types to be rebuilt functionally to update them, as most objects in DimensionalData are immutable.

x can be a AbstractDimArray, a Dimension, IndexMode or other custom types.

The arguments reuired are defined for the abstract type that has a rebuild method.

DimensionalData.modify — Function

modify(f, A::AbstractDimArray) => AbstractDimArray

Modify the parent data, rebuilding the AbstractDimArray wrapper without change. f must return a AbstractArray of the same size as the original.

DimensionalData.dimwise — Function

dimwise(f, A::AbstractDimArray{T,N}, B::AbstractDimArray{T2,M) => AbstractDimArray{T3,N}

Dimension-wise application of function f to A and B.

Arguments

-a: AbstractDimArray to broacast from, along dimensions not in b. -b: AbstractDimArray to broadcast from all diensions. Dimensions must be a subset of a.

This is like broadcasting over every slice of A if it is sliced by the dimensions of B, and storing the value in dest.

DimensionalData.dimwise! — Function

dimwise!(f, dest::AbstractDimArray{T1,N}, A::AbstractDimArray{T2,N}, B::AbstractDimArray) => dest

Dimension-wise application of function f.

Arguments

-dest: AbstractDimArray to update -a: AbstractDimArray to broacast from, along dimensions not in b. -b: AbstractDimArray to broadcast from all diensions. Dimensions must be a subset of a.

This is like broadcasting over every slice of A if it is sliced by the dimensions of B, and storing the value in dest.

DimensionalData.setdims — Function

setdims(A::AbstractArray, newdims) => AbstractArray
setdims(::Tuple, newdims) => Tuple{Vararg{<:Dimension,N}}
setdims(dim::Dimension, newdims) => Dimension

Replaces the first dim matching <: basetypeof(newdim) with newdim, and returns a new object or tuple with the dimension updated.

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

newdim: Tuple or single Dimension or dimension Type.

Example

using DimensionalData

A = DimArray(ones(10, 10), (X, Y(10:10:100)))
B = setdims(A, Y('a':'j'))
val(dims(B, Y))

# output

'a':1:'j'

DimensionalData.swapdims — Function

swapdims(x::T, newdims) => T
swapdims(dims::Tuple, newdims) => Tuple{Dimension}

Swap dimensions for the passed in dimensions, in the order passed.

Passing in the Dimension types rewraps the dimension index, keeping the index values and metadata, while constructed Dimension objectes replace the original dimension. nothing leaves the original dimension as-is.

Arguments

x: any object with a dims method or a Tuple of Dimension.

newdim: Tuple of Dimension or dimension Type.

Example

julia> using DimensionalData

julia> A = DimArray(ones(10, 10, 10), (X, Y, Z));

DimensionalData.reorderindex — Function

reorderindex(A, order::Union{Order,Dimension{<:Order},Tuple})

Reorder every dims index to order, or reorder index for the the given dimension(s) to the Order they wrap.

order can be an Order, a single Dimension or a Tuple of Dimension.

DimensionalData.reorderarray — Function

reorderarray(A, order::Union{Order,Dimension{<:Order},Tuple})

Reorder the array to order for every axis, or reorder array for the the given dimension(s) to the Order they wrap.

order can be an Order, a single Dimension or a Tuple of Dimension.

DimensionalData.reorderrelation — Function

reorderrelation(A, order::Union{Order,Dimension{<:Order},Tuple})

Reorder relation to order for every dimension, or reorder relation for the the given dimension(s) to the Order they wrap.

This will reverse the array, not the dimension index.

order can be an Order, a single Dimension or a Tuple of Dimension.

DimensionalData.reverseindex — Function

reverseindex(A; dims) => AbstractDimArray
reverseindex(dim::Dimension) => Dimension

Reverse the dimension index.

DimensionalData.reversearray — Function

reversearray(A; dims) => AbstractDimArray
reversearray(dim::Dimension) => Dimension

Reverse the array order, and update the dim to match.

Missing docstring.

Missing docstring for flipindex. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fliparray. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fliprelation. Check Documenter's build log for details.

Non-exported methods for developers

DimensionalData.dims2indices — Function

dims2indices(dim::Dimension, lookup, [emptyval=Colon()]) => NTuple{Union{Colon,AbstractArray,Int}}

Convert a Dimension or Selector lookup to indices, ranges or Colon.

DimensionalData.formatdims — Function

formatdims(A, dims) => Tuple{Vararg{<:Dimension,N}}

Format the passed-in dimension(s) dims to match the array A.

This means converting indexes of Tuple to LinRange, and running identify. Errors are also thrown if dims don't match the array dims or size.

If a IndexMode hasn't been specified, an mode is chosen based on the type and element type of the index:

DimensionalData.reducedims — Function

reducedims(x, dimstoreduce)

Replace the specified dimensions with an index of length 1. This is usually to match a new array size where an axis has been reduced with a method like mean or reduce to a length of 1, but the number of dimensions has not changed.

IndexMode traits are also updated to correspond to the change in cell step, sampling type and order.

DimensionalData.slicedims — Function

slicedims(x, I) => Tuple{Tuple,Tuple}

Slice the dimensions to match the axis values of the new array

All methods returns a tuple conatining two tuples: the new dimensions, and the reference dimensions. The ref dimensions are no longer used in the new struct but are useful to give context to plots.

Called at the array level the returned tuple will also include the previous reference dims attached to the array.

Arguments

x: An AbstractDimArray, Tuple of Dimension, or Dimension

I: A tuple of Integer, Colon or AbstractArray

DimensionalData.comparedims — Function

comparedims(A::AbstractDimArray...)
comparedims(A::Tuple...)
comparedims(a, b)

Check that dimensions or tuples of dimensions are the same, and return the first valid dimension. If AbstractDimArrays are passed as arguments their dimensions are compared.

Empty tuples and nothing dimension values are ignored, returning the Dimension value if it exists.

DimensionalData.identify — Function

identify(indexmode, index)

Identify an IndexMode or its fields from index content and existing IndexMode.