API

To use the functionality of DimensionalData in your module, dispatch on AbstractDimensionalArray and AbstractDimension.

Arrays:

DimensionalData.AbstractDimensionalArray — Type

Parent type for all dimensional arrays.

DimensionalData.DimensionalArray — Type

DimensionalArray(data, dims, refdims, name)

The main subtype of AbstractDimensionalArray. Maintains and updates its dimensions through transformations and moves dimensions to refdims after reducing operations (like e.g. mean).

Core types

Dimensions:

DimensionalData.Dimension — Type

Dimension is the abstract supertype of all dimension types.

Example concrete implementations are X, Y, Z, Ti (Time), and the arbirary Dim{:custom} dimension.

Dimensions label the axes of an AbstractDimesnionalArray, or other dimensional data. They may also provide an alternate index to lookup for each array axis.

Example:

using Dates
x = X(2:2:10)
y = Y(['a', 'b', 'c'])
ti = Ti(DateTime(2021, 1):Month(1):DateTime(2021, 12))

A = DimensionalArray(rand(3, 5, 12), (y, x, ti));

For simplicity, the same Dimension types are also used as wrappers in getindex, like:

x = A[X(2), Y(3)]

Dimension can also wrap Selectors.

x = A[X(Between(3, 4)), Y(At('b'))]

Dimension objects may have mode and metadata fields to track additional information about the data and the index, and their relationship.

DimensionalData.DependentDim — Type

Abstract supertype for Dependent dimensions. These will plot on the Y axis.

DimensionalData.IndependentDim — Type

Abstract supertype for independent dimensions. Thise will plot on the X axis.

DimensionalData.XDim — Type

Abstract parent type for all X dimensions.

DimensionalData.YDim — Type

Abstract parent type for all Y dimensions.

DimensionalData.ZDim — Type

Abstract parent type for all Z dimensions.

DimensionalData.TimeDim — Type

Abstract parent type for all time dimensions.

For an index with Interval sampling the locus will automatically be set to Start(), as a date/time index generally defines the start of a month, second etc, not the central point as is more common with spatial data. `

DimensionalData.X — Type

X Dimension. X <: XDim <: IndependentDim

Example:

x = X(2:2:10)

DimensionalData.Y — Type

Y Dimension. Y <: YDim <: DependentDim

Example:

y = Y(['a', 'b', 'c'])

DimensionalData.Z — Type

Z Dimension. Z <: ZDim <: Dimension

Example:

z = Z(10:10:100)

DimensionalData.Ti — Type

Time Dimension. Ti <: TimeDim <: IndependentDim

Time is already used by Dates, so we use Ti to avoid clashing.

Example:

ti = Ti(DateTime(2021, 1):Month(1):DateTime(2021, 12))

DimensionalData.ParametricDimension — Type

Dimensions with user-set type paremeters

DimensionalData.Dim — Type

Dim{X}(val, mode, metadata)
Dim{X}(val=:; [mode=Auto()], [metadata=nothing])

A generic dimension. For use when custom dims are required when loading data from a file. The sintax is ugly and verbose to use for indexing, ie Dim{:lat}(1:9) rather than Lat(1:9). This is the main reason they are not the only type of dimension availabile.

dim = Dim{:custom}(['a', 'b', 'c'])

DimensionalData.AnonDim — Type

AnonDim()

Anonymous dimension. Used when extra dimensions are created, such as during transpose of a vector.

DimensionalData.@dim — Macro

@dim typ [supertype=Dimension] [name=string(typ)] [shortname=string(typ)]

Macro to easily define specific dimensions.

Example:

@dim Lat "Lattitude" "lat"
@dim Lon XDim "Longitude"

Selectors

DimensionalData.Selector — Type

Selectors are wrappers that indicate that passed values are not the array indices, but values to be selected from the dimension index, such as DateTime objects for a Ti dimension.

DimensionalData.At — Type

At(x, atol, rtol)
At(x; atol=nothing, rtol=nothing)

Selector that exactly matches the value on the passed-in dimensions, or throws an error. For ranges and arrays, every intermediate value must match an existing value - not just the end points.

x can be any value or Vector of values.

atol and rtol are passed to isapprox. For Number rtol will be set to Base.rtoldefault, otherwise nothing, and wont be used.

DimensionalData.Near — Type

Near(x)

Selector that selects the nearest index. With Points this is simply the index nearest to the contained value, however with Intervals it is the interval center nearest to the contained value. This will be offset from the index value for Start and End loci.

DimensionalData.Between — Type

Between(a, b)

Selector that retreive all indices located between 2 values.

For Intervals the whole interval must be lie between the values. For Points the points must fall between the 2 values. These different sampling traits will often give different results with the same index and values - this is the intended behaviour.

DimensionalData.Contains — Type

Contains(x)

Selector that selects the interval the value is contained by. If the interval is not present in the index, an error will be thrown.

Can only be used for Intervals or Categorical.

Index Modes

DimensionalData.IndexMode — Type

Traits describing the mode of a dimension.

DimensionalData.Aligned — Type

Supertype for IndexMode where the index is aligned with the array axes. This is by far the most common case.

DimensionalData.AbstractSampled — Type

An IndexMode whos index is aligned with the array, and is independent of other dimensions.

DimensionalData.Sampled — Type

A concrete implementation of AbstractSampled. Can be used to represent points or intervals, with sampling of Points or Intervals.

Fields

order: Order indicating array and index order

span::Span: Span indicating Regular or Irregular size of intervals or distance between points

sampling::Sampling: Sampling of Intervals or Points (the default)

DimensionalData.AbstractCategorical — Type

IndexModes for dimensions where the values are categories.

DimensionalData.Categorical — Type

An IndexMode where the values are categories.

Fields

order: Order indicating array and index order.

Order will not be determined automatically for Categorical, it instead defaults to Unordered()

DimensionalData.Unaligned — Type

Supertype for IndexMode where the Dimension index is not aligned to the grid.

Indexing with an Unaligned dimension must provide all other Unaligned dimensions.

DimensionalData.Transformed — Type

IndexMode that uses an affine transformation to convert dimensions from dims(mode) to dims(array).

Fields

dims: a tuple containing dimenension types or symbols matching the order needed by the transform function.

DimensionalData.NoIndex — Type

IndexMode that is identical to the array axis.

DimensionalData.Auto — Type

Automatic IndexMode. Will be converted automatically to another IndexMode when possible.

Order of arrays and indices:

DimensionalData.Order — Type

Traits for the order of the array, index and the relation between them.

DimensionalData.Unordered — Type

Trait indicating that the array or dimension has no order.

DimensionalData.Ordered — Type

Container object for dimension and array ordering.

The default is Ordered(Forward(), Forward(), Forward())

All combinations of forward and reverse order for data and indices seem to occurr in real datasets, as strange as that seems. We cover these possibilities by specifying the order of both explicitly.

Knowing the order of indices is important for using methods like searchsortedfirst() to find indices in sorted lists of values. Knowing the order of the data is then required to map to the actual indices. It's also used to plot the data later - which always happens in smallest to largest order.

Base also defines Forward and Reverse, but they seem overly complicated for our purposes.

DimensionalData.AutoOrder — Type

Order will be found automatically where possible.

This will fail for all types without isless methods.

DimensionalData.UnknownOrder — Type

Order is not known and can't be determined.

DimensionalData.Forward — Type

Indicates that the array or dimension is in the normal forward order.

DimensionalData.Reverse — Type

Indicates that the array or dimension is in the reverse order. Selector lookup or plotting will be reversed.

Index modes for Sampled

Loci

DimensionalData.Locus — Type

Locii indicate the position of index values in cells.

Locii are often Start for time series, but often Center for spatial data.

DimensionalData.Center — Type

Indicates dimension index that matches the center coordinates/time/position.

DimensionalData.Start — Type

Indicates dimension index that matches the start coordinates/time/position.

DimensionalData.End — Type

Indicates dimension index that matches the end coordinates/time/position.

DimensionalData.AutoLocus — Type

Indicates dimension where the index position is not known.

Span

DimensionalData.Span — Type

Mode defining the type of interval used in a InervalSampling index.

DimensionalData.Regular — Type

Intervalss have regular size. This is passed to the constructor, although these are normally build automatically.

DimensionalData.Irregular — Type

Irregular have irrigular size. To enable bounds tracking and accuract selectors, the starting bounds must be provided as a 2 tuple, or 2 arguments.

DimensionalData.AutoSpan — Type

Span will be guessed and replaced by a constructor.

Sampling

DimensionalData.Sampling — Type

Indicates the sampling method used by the index.

DimensionalData.Points — Type

Sampling mode where single samples at exact points.

DimensionalData.Intervals — Type

Sampling mode where samples are the mean (or similar) value over an interval.

Methods

Getting basic info

These useful functions for obtaining information from your dimensional data:

DimensionalData.bounds — Function

bounds(x, [dims])

Return the bounds of all dimensions of an object, of a specific dimension, or of a tuple of dimensions.

Returns a length 2 Tuple in ascending order.

DimensionalData.data — Function

data(x)

Return the data wrapped by the dimentional array. This may not be the same as Base.parent, as it should never include data outside the bounds of the dimensions.

In a disk based AbstractDimensionalArray, data may need to load data from disk.

DimensionalData.dimnum — Function

dimnum(x, lookup)

Get the number(s) of Dimension(s) as ordered in the dimensions of an object.

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

lookup: Tuple or single Dimension or dimension Type.

The return type will be a Tuple of Int or a single Int, depending on wether lookup is a Tuple or single Dimension.

Example

julia> A = DimensionalArray(rand(10, 10, 10), (X, Y, Z));

julia> dimnum(A, Z)
3

DimensionalData.dims — Function

dims(x)

Return a tuple of Dimensions for an object, in the order that matches the axes or columns etc. of the underlying data.

DimensionalData.hasdim — Function

hasdim(x, lookup)

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

lookup: Tuple or single Dimension or dimension Type.

Check if an object or tuple contains an Dimension, or a tuple of dimensions.

Example

julia> A = DimensionalArray(rand(10, 10, 10), (X, Y, Z));

julia> hasdim(A, X)
true

julia> hasdim(A, Ti)
false

DimensionalData.label — Function

label(x)

Get a plot label for data or a dimension. This will include the name and units if they exist, and anything else that should be shown on a plot.

DimensionalData.metadata — Function

metadata(x)

Return the metadata of a dimension or data object.

DimensionalData.name — Function

name(x)

Get the name of data or a dimension.

DimensionalData.rebuild — Function

rebuild(x, args...)
rebuild(x; kwargs...)

Rebuild an object struct with updated values.

DimensionalData.refdims — Function

refdims(x)

Reference dimensions for an array that is a slice or view of another array with more dimensions.

slicedims(a, dims) returns a tuple containing the current new dimensions and the new reference dimensions. Refdims can be stored in a field or disgarded, as it is mostly to give context to plots. Ignoring refdims will simply leave some captions empty.

DimensionalData.shortname — Function

shortname(x)

Get the short name of array data or a dimension.

This may be a shorter version more suitable for small labels than name, but it may also be identical to name.

DimensionalData.units — Function

units(x)

Return the units of a dimension. This could be a string, a unitful unit, or nothing.

Units do not have a field, and may or may not be included in metadata. This method is to facilitate use in labels and plots when units are available, not a guarantee that they will be.

DimensionalData.val — Function

val(x)

Return the contained value of a wrapper object, otherwise just returns the object.

And some utility methods:

Missing docstring.

Missing docstring for setdim. Check Documenter's build log for details.

DimensionalData.swapdims — Function

swapdims(x, newdims)

Swap the dimension for the passed in dimensions. Dimension wrapper types rewrap the original dimension, keeping the values and metadata. Dimension instances replace the original dimension, and nothing leaves the original dimension as-is.

Arguments

x: any object with a dims method, a Tuple of Dimension or a single Dimension.

newdim: Tuple or single Dimension or dimension Type.

Example

julia> A = DimensionalArray(rand(10, 10, 10), (X, Y, Z));

julia> B = swapdims(A, (Z, Dim{:custom}, Ti));

julia> dimnum(B, Ti)
3

Low-level methods

DimensionalData.dims2indices — Function

dims2indices(dim::Dimension, lookup, [emptyval=Colon()])

Convert a Dimension or Selector lookup to indices, ranges or Colon.

dims2indices(A, lookup, [emptyval=Colon()])

Convert Dimension or Selector to regular indices for any object with a dims method, usually an array.

dims2indices(dims, lookup, [emptyval=Colon()])

Convert Dimension or Selector to regular indices for dims - a Tuple of Dimension. lookup can be a Tuple or a single object.

DimensionalData.formatdims — Function

formatdims(A, dims)

Format the passed-in dimension(s).

Mostily this means converting indexes of tuples and UnitRanges to LinRange, which is easier to handle internally. Errors are also thrown if dims don't match the array dims or size.

If a IndexMode hasn't been specified, an mode is chosen based on the type and element type of the index:

DimensionalData.reducedims — Function

reducedims(x, dimstoreduce)

Replace the specified dimensions with an index of length 1. This is usually to match a new array size where an axis has been reduced with a method like mean or reduce to a length of 1, but the number of dimensions has not changed.

IndexMode traits are also updated to correspond to the change in cell step, sampling type and order.

DimensionalData.slicedims — Function

Slice the dimensions to match the axis values of the new array

All methods returns a tuple conatining two tuples: the new dimensions, and the reference dimensions. The ref dimensions are no longer used in the new struct but are useful to give context to plots.

Called at the array level the returned tuple will also include the previous reference dims attached to the array.

DimensionalData.comparedims — Function

comparedims(a, b)

Check that dimensions or tuples of dimensions are the same. Empty tuples are allowed

DimensionalData.identify — Function

identify(indexmode, index)

Identify an IndexMode or its fields from index content and existing IndexMode.

DimensionalData.DimensionalStyle — Type

DimensionalStyle{S}

This is a BroadcastStyle for AbstractAbstractDimensionalArray's It preserves the dimension names. S should be the BroadcastStyle of the wrapped type.

Copied from NamedDims.jl (thanks @oxinabox).